### 2021-11-26 #3954 @whedon
[PRE REVIEW]: ViMMS 2.0: A framework to develop, test and optimise 2 fragmentation strategies in LC-MS metabolomics
**Submitting author:** @joewandy (<a href="http://orcid.org/0000-0002-3068-4664">Joe Wandy</a>)
**Repository:** <a href="https://github.com/glasgowcompbio/vimms/" target ="_blank">https://github.com/glasgowcompbio/vimms/</a>
**Version:** v2.0
