### 2021-11-26 #3954 @Kevin-Mattheus-Moerman
@joewandy could you check those missing DOI's :point_up:  you can call `@whedon generate pdf` here to update the paper or call `@whedon check references` to check those DOI's again. Thanks.
