### 2021-11-29 a700141
Added missing DOIs in paper.bib
