+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={316--320},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nbt.4101}
 }
 
 @ARTICLE{smith2014proteomics,
@@ -17,7 +18,8 @@
   number={7},
   pages={1--14},
   year={2014},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi = {10.1186/1471-2105-15-s7-s9}
 }
 
 @ARTICLE{Wandy2019-ok,
@@ -35,7 +37,8 @@
   keywords  = "data-dependent acquisition (DDA); fragmentation (MS/MS); in
                silico; liquid chromatography--mass spectrometry (LC/MS);
                simulator",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1101/744227}
 }
 
 @article{davies21_smartroi,
@@ -46,7 +49,8 @@
   number={14},
   pages={5676--5683},
   year={2021},
-  publisher={American Chemical Society}
+  publisher={American Chemical Society},
+  doi={10.1021/acs.analchem.0c03895}
 }
 
 @article{smith2015jamss,
@@ -57,7 +61,8 @@
   number={5},
   pages={791--793},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btu729}
 }
 
 @article{noyce2013mspire,
@@ -68,7 +73,8 @@
   number={12},
   pages={5742--5749},
   year={2013},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/pr400727e}
 }
 
 @article{goldfarb2016msacquisitionsimulator,
@@ -79,7 +85,8 @@
   number={8},
   pages={1269--1271},
   year={2016},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btv745}
 }
 
 @article{kosters2021smiter,
@@ -90,7 +97,8 @@
   number={3},
   pages={396},
   year={2021},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={10.3390/genes12030396}
 }
 
 @article{wang2015openms,
@@ -101,7 +109,8 @@
   number={1},
   pages={1--6},
   year={2015},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s12859-015-0540-1}
 }
 
 @MISC{Thermo_Fisher_Scientific_undated-ny,
