### 2021-11-29 #3954 @joewandy
> @joewandy could you check those missing DOI's ☝️ you can call `@whedon generate pdf` here to update the paper or call `@whedon check references` to check those DOI's again. Thanks.

Thanks for letting me know. I've fixed those missing DOIs and regenerated the paper.
