### 2021-11-29 #3954 @joewandy
> 👋🏻 @joewandy - I will handle your submission and start by looking for suitable reviewers. If you have suggestions, please let me know here (without tagging them with `@`).

Hello i'd like to suggest the following people as reviewers as I'm aware they have worked on similar things in the past:

MKoesters -- worked on [SMITER](https://www.mdpi.com/2073-4425/12/3/396)
lfnothias -- worked on [MS2Planner](https://academic.oup.com/bioinformatics/article/37/Supplement_1/i231/6319686)
florian-huber -- did a lot of work on deep learning and mass spec data
