### 2021-11-29 e60ee7b
Small fixes to make linting happy
