### 2021-12-01 #3954 @csoneson
👋🏻   @MKoesters, @florian-huber, @jorainer - would two of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> ViMMS 2.0: A framework to develop, test and optimise 2 fragmentation strategies in LC-MS metabolomics

You can find more information about JOSS and the review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) - of course, happy to answer any questions as well.
