### 2021-12-02 3872797
Correctly compute WeightedDEW weight (issue #226)
