### 2021-12-02 0bc4a8b
Method added to turn RoIAligner into plottable boxes. Some small rearrangements to make use easier.
