### 2021-12-06 #3954 @jorainer
Sorry for my late reply - super busy at present. Thus I'll also not have time to review at present - besides, I'm not a python developer, so I can't provide usefull feedback on the actual implementation.
