### 2021-12-06 #3954 @florian-huber
I could review the manuscript!

One disclaimer though: I did collaborate with one of the authors (Simon Rogers), including a publication that was finished >1 year ago. The collaboration was on a very different project and I was not involved in any way in the project presented here. So, I believe that I can make an impartial assessment of the work, but that's up to you to decide @csoneson .
