### 2021-12-06 #3954 @csoneson
👋🏻 @hechth, @jspaezp - would you be interested in reviewing this submission for JOSS?

> ViMMS 2.0: A framework to develop, test and optimise 2 fragmentation strategies in LC-MS metabolomics
