### 2021-12-07 #3954 @csoneson
@hechth - it would be great if you could come back to me before the end of the week so that, if needed, I can look for another reviewer without too much delay. If you would like some more information about reviewing for JOSS, including the review criteria and checklist, see [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) - of course, I'm happy to answer any questions as well. Thanks!
