### 2021-12-13 #3990 @whedon
[REVIEW]: ViMMS 2.0: A framework to develop, test and optimise 2 fragmentation strategies in LC-MS metabolomics
**Submitting author:** @joewandy (<a href="http://orcid.org/0000-0002-3068-4664">Joe Wandy</a>)
**Repository:** https://github.com/glasgowcompbio/vimms/
**Branch with paper.md** (empty if default branch):
**Version:** V2.0.5
