+++ requirements.txt
@@ -58,7 +58,7 @@ jupyterlab-server==2.8.1
 keyring==23.2.1
 kiwisolver==1.3.2
 loguru==0.5.3
-lxml==4.6.3
+lxml==4.6.5
 Markdown==3.3.4
 MarkupSafe==2.0.1
 mass-spec-utils==0.0.12
