### 2021-12-16 9c32b2b
Support for plotting fragmentation events and boxes of different colours added. Various fixes.
