+++ vimms/Box_Visualise.py
@@ -84,10 +84,16 @@ class InterpolationMap(ColourMap):
         intervals = [minm + i * (maxm - minm) / (len(self.colours) - 1) for i in range(len(self.colours))]
     
         def get_colour(box):
-            if(math.isclose(minm, maxm)): return self.colours[0]
-            i = next(i-1 for i, threshold in enumerate(intervals) if threshold >= key(box))
-            weight = (key(box) - intervals[i]) / (intervals[i + 1] - intervals[i])
-            return self.colours[i].interpolate([self.colours[i+1]], weights=[1-weight, weight])
+            if(math.isclose(minm, maxm)): 
+                return self.colours[0]
+            elif(key(box) < intervals[1]): 
+                return self.colours[0]
+            elif(key(box) > intervals[-1]): 
+                return self.colours[-1]
+            else:
+                i = next(i-1 for i, threshold in enumerate(intervals) if threshold >= key(box))
+                weight = (key(box) - intervals[i]) / (intervals[i + 1] - intervals[i])
+                return self.colours[i].interpolate([self.colours[i+1]], weights=[1-weight, weight])
         
         return ((b, get_colour(b)) for b in boxes)
         
@@ -138,9 +144,8 @@ class AutoColourMap(ColourMap):
 
 def path_or_mzml(mzml):    
     try:
-        path = Path(mzml)
-        mzml = MZMLFile(path)
-    except:
+        mzml = MZMLFile(mzml)
+    except FileNotFoundError:
         if(not type(mzml) == MZMLFile):
             raise NotImplementedError("Didn't recognise the MZMLFile!")
     return mzml
@@ -149,7 +154,8 @@ class PlotPoints():
     def __init__(self, ms1_points, ms2s=None, markers={}):
         self.ms1_points, self.ms2s = ms1_points, ms2s
         self.markers = markers
-        self.active = np.ones_like(ms1_points)
+        self.active_ms1 = np.ones_like(ms1_points)
+        self.active_ms2 = np.ones_like(ms2s)
         
     @staticmethod
     def from_mzml(mzml):
@@ -159,20 +165,27 @@ class PlotPoints():
             if(s.ms_level in scan_dict): scan_dict[s.ms_level].append(s)
         scan_dict[1] = sorted(scan_dict[1], key=lambda s: s.rt_in_seconds)
         ms1_points = np.array([[s.rt_in_seconds, mz, intensity] for s in scan_dict[1] for mz, intensity in s.peaks])
-        return PlotPoints(ms1_points, [(s.rt_in_seconds, s.precursor_mz) for s in scan_dict[2]])
+        ms2s = np.array([[s.rt_in_seconds, s.precursor_mz, s.get_max_intensity().intensity] for s in scan_dict[2]])
+        return PlotPoints(ms1_points, ms2s=ms2s)
+        
+    def bound_points(self, pts, min_rt=None, max_rt=None, min_mz=None, max_mz=None):
+        all_true = np.array(np.ones_like(pts[:, 0]), dtype=np.bool)
+        select_rt = (all_true if (min_rt is None) else (pts[:, 0] >= min_rt)) & (all_true if (max_rt is None) else (pts[:, 0] <= max_rt))
+        select_mz = (all_true if (min_mz is None) else (pts[:, 1] >= min_mz)) & (all_true if (max_mz is None) else (pts[:, 1] <= max_mz))
+        return (select_rt & select_mz)
         
[…]
