### 2021-12-16 9ab95e4
Added support for plotting any boxes within range of the visible bounds of the canvas when plotting an individual box as focus, and support for manually setting visible area in this case. Fixed some breaking changes.
