### 2021-12-16 8401002
Added more ways to create PlotBoxes. Small fixes.
