### 2022-01-01 glasgowcompbio/vimms#230 @jspaezp
JOSS Review (jspaezp)
Review checklist for @jspaezp for JOSS https://github.com/openjournals/joss-reviews/issues/3990

- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

Code of Conduct

- [x]  I confirm that I read and will adhere to the JOSS code of conduct.

General checks

- [x] Repository: Is the source code for this software available at the repository url?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [x] Contribution and authorship: Has the submitting author (@joewandy) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [x]  Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

    - The installation mentions `ViMMS requires Python 3+. Unfortunately it is not compatible with Python 2. You can install a stable version of ViMMS using pip or pipenv.`. I don't think this line is needed, since Py2 has reached EOL, therefore it is reasonable to assume that Python 3.6+ will be used.

    - [x] there is mention of `git clone https://github.com/sdrogers/vimms.git`, please update to the new repository url (glasgowcompbio/vimms)
    - [x] https://github.com/glasgowcompbio/vimms/blob/25afd804479ddb58cd30ed93287428872b1a8adc/setup.py#L15 same here
    - [x] `RoiToChemicalCreator` is imported in the examples, which raises `ImportError: cannot import name 'RoiToChemicalCreator' from 'vimms.Roi' (~/.local/share/virtualenvs/vimms-j1m4gYsN/lib/python3.9/site-packages/vimms/Roi.py)` please update the examples or add annotations to the directories specifying which ones are meant to run from the current master branch (and ideally if they were meant to run in the past, what commit that was)
    - The long form documentation is exemplary good (examples/demos) but the API and code level documentation (readthedocs/docstrings) is not great, I believe that adding much better information on the parameters is required.
    - [x] Parameter docstrings
- [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
    - [x] Please add a README to `https://github.com/glasgowcompbio/vimms/tree/master/tests/fixtures/` with instructions on how to generate those files. (and ideally remove those files so they are generated and cached during the testing process).
    - [ ] I believe a codecov report should be added to the CI pipeline
- [x] Community guidelines: Are there clear guidelines for third parties wishing to
    1) [x] Contribute to the software
        - **I would appreciate if the guideline included any specifics on what linter/checker/testing framework to use during the code contribution. (not critical)**
        -  [x] Is there any intention of making the code flake8 compliant? I noticed the CI checks for it but based on the number of "violations" I don't see the point of having it (https://github.com/glasgowcompbio/vimms/runs/4901527800?check_suite_focus=true)
    2) [x] Report issues or problems with the software
    3) [x] Seek support

1. On the README there is the reference to: `Run Jupyter ($ jupyter lab) so you could run the example notebooks below.`, please add a link to the notebooks being referenced.

- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is - [] designed to solve and who the target audience is?
- [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Code Review:

1. https://github.com/glasgowcompbio/vimms/blob/387279719b7769658b267846aee3dfa8ee6dfa0c/vimms/Common.py#L278-L281
    1. Should a warning be thrown if the length of the path is "." instead of just silently ignoring the command?
    2.  I believe that check is redundant with the behavior of pathlib.Path.mkdir
2. https://github.com/glasgowcompbio/vimms/blob/387279719b7769658b267846aee3dfa8ee6dfa0c/vimms/Common.py#L305-L316
    1. I believe this function should throw an error when the file cannot be open instead of a warn and returning None
3.  https://github.com/glasgowcompbio/vimms/blob/387279719b7769658b267846aee3dfa8ee6dfa0c/vimms/Common.py#L319-L327
    1. Should an assertion be added to make sure the two iterables have the same length? As it is right now it will silently give a wrong answer when `len(intensities) > len(rts)` but break when `len(rts) > len(intensities)`
4. https://github.com/glasgowcompbio/vimms/blob/387279719b7769658b267846aee3dfa8ee6dfa0c/vimms/Common.py#L16
    1. I suggest you import tqdm as `from tqdm.auto import tqdm`, since it handles better progress bars in different plattaforms (automatically selects a notebook/terminal bar)
5. I think the object nomenclature should be a little more clear when saving pickled objects. Since the `save_obj` function saves files as gzipped pickled objects, having the extension be `.p.gz` could be more appropriate than `.p`
6. Empty `vimms/demo/03. MultiSampleMethods/`
7. Is there any reason `vimms/notebook` is in that location? (inside the package itselt)
8. https://github.com/glasgowcompbio/vimms/blob/94da3c7654429a6e2e36383666b002194d951cc5/vimms/Controller/base.py#L55 I believe this should be an instance of `from abc import ABC` and have the methods required for the functioning of the class be abstractmethods  `@abc.abstractmethod`. So errors are raised during instantiation and not during calling of the method.
    9.  https://github.com/glasgowcompbio/vimms/blob/69afe3af55e98138bdb0f95a09302b7117eb18ff/vimms/Agent.py#L12 same here
9. https://github.com/glasgowcompbio/vimms/blob/2162c09203fa8899015ff658bf57642259f857bb/vimms/Controller/vinny_test.py#L200 Raising not implemented errors should be written as `raise NotImplementedError`
10. `vimms/old_unused_experimental` please remove from the main repository (potentially generate a branch so they are not "lost")
11. https://github.com/glasgowcompbio/vimms/blob/453e989f4e26c50c0e20a1e887c53e012bd040ad/vimms/Chemicals.py#L37 This definition should probably come from vimms.Common
