### 2022-01-10 25afd80
Small fix for pickling problem
