### 2022-01-11 53e8de1
Typechecking on __eq__ in Matching.py objects
