+++ requirements.txt
@@ -93,7 +93,7 @@ pbr==5.6.0
 pep517==0.11.0
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.3.2
+Pillow==9.0.0
 pip-shims==0.5.3
 pipenv-setup==3.1.1
 pipfile==0.0.2
