### 2022-01-14 09c0188
Fix for Python < 3.8: start is a positional argument in sum.
