### 2022-01-14 99bf61b
Small fix to be able to plot, and corrected shape of indexing array.
