### 2022-01-17 78ed208
Allow environment to dump debug info (as pickled)
