### 2022-01-17 28ec68b
Small fixes for SmartRoi in RoiBuilder
