### 2022-01-17 6f43089
Experimenting with intensity diff
