### 2022-01-18 56e07af
Score changes for non_overlap_smartroi controller
