+++ vimms/Controller/box.py
@@ -53,24 +53,21 @@ class NonOverlapController(GridController):
                         r in self.roi_builder.live_roi])
         return non_overlaps
 
+    def _scale(self, scores):
+        if len(scores) > 0 and max(scores) > 0:
+            scores = scores / max(scores)
+        return scores
+
     def _get_scores(self):
         dda_scores = self._get_dda_scores()
-        if len(dda_scores) > 0 and max(dda_scores) > 0:
-            dda_scores = dda_scores / max(dda_scores)
-
         non_overlaps = self._overlap_scores()
-        final_scores = dda_scores + non_overlaps
         if self.roi_builder.roi_type == RoiBuilder.ROI_TYPE_SMART:
+            w1, w2, w3 = 1, 1, 1
             smartroi_scores = self._smartroi_filter()
-            if len(smartroi_scores) > 0 and max(smartroi_scores) > 0:
-                smartroi_scores = smartroi_scores / max(smartroi_scores)
-            final_scores = final_scores + smartroi_scores
-
-        for i in range(len(non_overlaps)):
-            non = non_overlaps[i]
-            if non < 0.5:
-                final_scores[i] = 0
-
+            dda_scores = self._scale(dda_scores)
+            final_scores = (w1*dda_scores) + (w2*smartroi_scores) + (w3*non_overlaps)
+        else:
+            final_scores = dda_scores * non_overlaps
         return self._get_top_N_scores(final_scores)
 
 
