### 2022-01-19 953799c
First pass modification of RoI aligner code. More changes to follow.
