+++ vimms/Roi.py
@@ -1,5 +1,7 @@
+import copy
 import bisect
 import sys
+from statistics import mean
 from collections import OrderedDict
 
 import pandas as pd
@@ -457,34 +459,25 @@ class RoiAligner(object):
                  min_rt_width=0.01,
                  min_mz_width=0.01,
                  n_categories=1):
-        self.peaksets = []
-        self.mz_tolerance_absolute = mz_tolerance_absolute
-        self.mz_tolerance_ppm = mz_tolerance_ppm
-        self.rt_tolerance = rt_tolerance
-        self.mz_weight = 75
-        self.rt_weight = 25
-        self.files_loaded = []
-        self.mz_column_pos = mz_column_pos
-        self.rt_column_pos = rt_column_pos
-        self.intensity_column_pos = intensity_column_pos
-        self.sample_names = []
-        self.sample_types = []
-        self.min_rt_width = min_rt_width
-        self.min_mz_width = min_mz_width
-        self.peaksets2boxes = {}
-        self.peaksets2fragintensities = {}
-        self.addition_method = None
+
+        self.mz_tolerance_absolute, self.mz_tolerance_ppm, self.rt_tolerance = mz_tolerance_absolute, mz_tolerance_ppm, rt_tolerance
+        self.mz_column_pos, self.rt_column_pos, self.intensity_column_pos = mz_column_pos, rt_column_pos, intensity_column_pos
+        self.min_rt_width, self.min_mz_width = min_rt_width, min_mz_width
         self.n_categories = n_categories
-        self.list_of_boxes = []
+
+        self.peaksets, self.files_loaded, self.list_of_boxes = [], [], []
+        self.sample_names, self.sample_types = [], []
+        self.mz_weight, self.rt_weight = 75, 25
+        self.peaksets2boxes, self.peaksets2fragintensities = {}, {}
+        self.addition_method = None
 
     def add_sample(self, rois, sample_name, sample_type=None, rt_shifts=None, mz_shifts=None):
         self.sample_names.append(sample_name)
         self.sample_types.append(sample_type)
-        these_peaks = []
-        frag_intensities = []
-        temp_boxes = []
+        
+        these_peaks, frag_intensities, temp_boxes = [], [], []
         for i, roi in enumerate(rois):
-            source_id = sample_name + '_' + str(i)
+            source_id = f"{sample_name}_{i}"
             peak_mz = roi.get_mean_mz()
             peak_rt = roi.estimate_apex()
             peak_intensity = roi.get_max_intensity()
@@ -496,37 +489,41 @@ class RoiAligner(object):
 
         # do alignment, adding the peaks and boxes, and recalculating max frag intensity
         self._align(these_peaks, temp_boxes, frag_intensities, sample_name)
+        
+    @staticmethod
+    def load_boxes(peak_file, picking_method):
+        if picking_method == "mzmine":
+            boxes = load_picked_boxes(peak_file)
+        elif picking_method == "peakonly":
+            boxes = load_peakonly_boxes(peak_file)  # not tested
+        elif picking_method == "xcms":
+            boxes = load_xcms_boxes(peak_file)  # not tested
+        else:
+            raise NotImplementedError(f"Picking method \"{picking_method}\" not recognised!")
+        return boxes
 
-    def add_picked_peaks(self, mzml_file, peak_file, sample_name, picking_method='mzmine', sample_type=None,
+    def add_picked_peaks(self, mzml_file, peak_file, sample_name, picking_method="mzmine", sample_type=None,
                          half_isolation_window=0, allow_last_overlap=False, rt_shifts=None, mz_shifts=None):
         self.sample_names.append(sample_name)
         self.sample_types.append(sample_type)
-        these_peaks = []
-        frag_intensities = []
-        # load boxes
[…]
