### 2022-01-19 b4c1a84
Code tidy up
