### 2022-01-20 efc1d16
non_overlap + weighteddew fixes.
