### 2022-01-21 #3990 @jspaezp
Hello @csoneson, Indeed I have been working on the review here: https://github.com/glasgowcompbio/vimms/issues/230 overall the project is a lot larger than I expected and It has been taking longer than I though. I Should be able to finish the review by this weekend. (overall it seems to comply with all the criteria for acceptance, but there are a couple of things I would like to suggest as "code review")
