### 2022-01-21 99e6176
Fixes for intensity non-overlap too
