+++ vimms/Controller/box.py
@@ -1,9 +1,8 @@
-from abc import abstractmethod
 from copy import deepcopy
 
 import numpy as np
 
-from vimms.Common import ROI_EXCLUSION_DEW
+from vimms.Common import ROI_EXCLUSION_DEW, GRID_CONTROLLER_SCORING_PARAMS
 from vimms.Controller.roi import RoiController, RoiBuilder
 
 
@@ -12,12 +11,10 @@ class GridController(RoiController):
     def __init__(self, ionisation_mode, isolation_width, mz_tol, min_ms1_intensity, min_roi_intensity,
                  min_roi_length, N, grid, rt_tol=10, min_roi_length_for_fragmentation=1, length_units="scans",
                  ms1_shift=0, min_rt_width=0.01, min_mz_width=0.01,
-                 params=None, register_all_roi=False,
+                 params=None, register_all_roi=False, scoring_params=GRID_CONTROLLER_SCORING_PARAMS,
                  roi_type=RoiBuilder.ROI_TYPE_NORMAL, reset_length_seconds=1e6,  # smartroi parameters
                  intensity_increase_factor=10, drop_perc=0.1 / 100,  # smartroi parameters
-                 exclusion_method=ROI_EXCLUSION_DEW, exclusion_t_0=None,  # weighted dew parameters
-                 dda_weight=1, overlap_weight=1, smartroi_weight=1
-                 ):
+                 exclusion_method=ROI_EXCLUSION_DEW, exclusion_t_0=None):  # weighted dew parameters
         super().__init__(
             ionisation_mode, isolation_width, mz_tol, min_ms1_intensity, min_roi_intensity,
             min_roi_length, N, rt_tol=rt_tol, min_roi_length_for_fragmentation=min_roi_length_for_fragmentation,
@@ -35,17 +32,42 @@ class GridController(RoiController):
         self.min_rt_width, self.min_mz_width = min_rt_width, min_mz_width
         self.grid = grid  # helps us understand previous RoIs
         self.register_all_roi = register_all_roi
-        self.dda_weight = dda_weight
-        self.overlap_weight = overlap_weight
-        self.smartroi_weight = smartroi_weight
+
+        self.scoring_params = scoring_params
+        self.dda_weight = scoring_params['dda_weight']
+        self.overlap_weight = scoring_params['overlap_weight']
+        self.smartroi_weight = scoring_params['smartroi_weight']
+        self.smartroi_score_add = scoring_params['smartroi_score_add']
 
     def update_state_after_scan(self, scan):
         super().update_state_after_scan(scan)
         self.grid.send_training_data(scan)
 
-    @abstractmethod
+    def _scale(self, scores):
+        if len(scores) > 0 and max(scores) > 0:
+            scores = scores / max(scores)
+        return scores
+
     def _get_scores(self):
-        pass
+        non_overlaps = self._overlap_scores()
+        if self.roi_builder.roi_type == RoiBuilder.ROI_TYPE_SMART:  # smart ROI scoring
+            smartroi_scores = self._smartroi_filter()
+            dda_scores = self._log_roi_intensities() * self._min_intensity_filter()
+
+            if self.smartroi_score_add:  # add the scores
+                dda_scores = self._scale(dda_scores)
+                final_scores = (self.dda_weight * dda_scores) + (self.smartroi_weight * smartroi_scores) + (
+                        self.overlap_weight * non_overlaps)
+
+            else:  # multiply them, this wouldn't work well because a lot of the smartroi scores are 0s
+                final_scores = dda_scores * smartroi_scores * non_overlaps
+
+        else:  # normal ROI
+            dda_scores = self._get_dda_scores()
+            final_scores = dda_scores * non_overlaps
+
+        # print(final_scores)
[…]
