+++ requirements.txt
@@ -37,7 +37,7 @@ iniconfig==1.1.1
 intervaltree==3.1.0
 ipykernel==6.4.1
 ipyparallel==7.0.1
-ipython==7.27.0
+ipython==7.31.1
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
 jedi==0.18.0
