### 2022-01-24 fb971ef
Another refactor of the internals of _align. path_or_mzml moved to Common - sometimes you might want to construct the MZMLFile object only once before passing it, so this lets paths or an MZMLFile object be used. This could be moved to the constructor later though.
