### 2022-01-28 #3990 @jspaezp
@joewandy I believe that we can start discussing/addressing several of the issues raised in my thread. My general perception is that there is really good code quality and long form documentation. Nonetheless there is a fair amount of technical debt in the form of unused sub-directories and in-code documentation (docstrings/parameter documentation), which is natural for a project of this size/age. In general it is also unclear in most of the codebase what API should be user-facing and what should be internal, there seems to be a pretty arbitrary usage of the hidden methods/classes.
@csoneson I cannot edit the checklist in this issue, there is a copy on the issue i created in the evaluated repository.
