+++ vimms/Chemicals.py
@@ -386,8 +386,7 @@ class ChemicalMixtureFromMZML(object):
                               mz_units=self.roi_params.mz_units, \
                               min_length=self.roi_params.min_length, min_intensity=self.roi_params.min_intensity, \
                               start_rt=self.roi_params.start_rt, stop_rt=self.roi_params.stop_rt,
-                              length_units=self.roi_params.length_units, \
-                              ms_level=self.roi_params.ms_level, skip=self.roi_params.skip)
+                              length_units=self.roi_params.length_units)
         logger.debug("Extracted {} good ROIs from {}".format(len(good), self.mzml_file_name))
         return good
 
