### 2022-01-28 c92716f
Moved ROI Builder to Roi.py. Initial code tidy up.
