+++ vimms/Roi.py
@@ -1,6 +1,10 @@
+"""
+Provides implementation of Regions-of-Interest (ROI) objects that are used for real-time ROI tracking
+in various controller. Additionally, ROIs can also be loaded from an mzML file and converted into
+Chemical objects for simulation input.
+"""
 import bisect
 import sys
-from collections import OrderedDict
 
 import pandas as pd
 import pylab as plt
@@ -11,25 +15,7 @@ from scipy.stats import pearsonr
 
 from vimms.Box import GenericBox
 from vimms.Chromatograms import EmpiricalChromatogram
-from vimms.Common import PROTON_MASS
 from vimms.Evaluation import *
-
-POS_TRANSFORMATIONS = OrderedDict()
-POS_TRANSFORMATIONS['M+H'] = lambda mz: (mz + PROTON_MASS)
-POS_TRANSFORMATIONS['[M+ACN]+H'] = lambda mz: (mz + 42.033823)
-POS_TRANSFORMATIONS['[M+CH3OH]+H'] = lambda mz: (mz + 33.033489)
-POS_TRANSFORMATIONS['[M+NH3]+H'] = lambda mz: (mz + 18.033823)
-POS_TRANSFORMATIONS['M+Na'] = lambda mz: (mz + 22.989218)
-POS_TRANSFORMATIONS['M+K'] = lambda mz: (mz + 38.963158)
-POS_TRANSFORMATIONS['M+2Na-H'] = lambda mz: (mz + 44.971160)
-POS_TRANSFORMATIONS['M+ACN+Na'] = lambda mz: (mz + 64.015765)
-POS_TRANSFORMATIONS['M+2Na-H'] = lambda mz: (mz + 44.971160)
-POS_TRANSFORMATIONS['M+2K+H'] = lambda mz: (mz + 76.919040)
-POS_TRANSFORMATIONS['[M+DMSO]+H'] = lambda mz: (mz + 79.02122)
-POS_TRANSFORMATIONS['[M+2ACN]+H'] = lambda mz: (mz + 83.060370)
-POS_TRANSFORMATIONS['2M+H'] = lambda mz: (mz * 2) + 1.007276
-POS_TRANSFORMATIONS['M+ACN+Na'] = lambda mz: (mz + 64.015765)
-POS_TRANSFORMATIONS['2M+NH4'] = lambda mz: (mz * 2) + 18.033823
 
 
 class Roi(object):
@@ -39,23 +25,21 @@ class Roi(object):
     maintains 3 lists -- mz, rt and intensity. When a new point (mz,rt,intensity) is added, it updates the list and the
     mean mz which is required.
     """
+
     def __init__(self, mz, rt, intensity, id=None):
+        """
+        Constructs a new ROI
+        :param mz: the initial m/z of this ROI. Can be either a single value or a list of values.
+        :param rt: the initial rt  of this ROI. Can be either a single value or a list of values.
+        :param intensity: the initial intensity  of this ROI. Can be either a single value or a list of values.
+        """
         self.id = id
         self.fragmentation_events = []
         self.fragmentation_intensities = []
         self.max_fragmentation_intensity = 0.0
-        if type(mz) == list:
-            self.mz_list = mz
-        else:
-            self.mz_list = [mz]
-        if type(rt) == list:
-            self.rt_list = rt
-        else:
-            self.rt_list = [rt]
-        if type(intensity) == list:
-            self.intensity_list = intensity
-        else:
-            self.intensity_list = [intensity]
+        self.mz_list = self._to_list(mz)
+        self.rt_list = self._to_list(rt)
+        self.intensity_list = self._to_list(intensity)
         self.n = len(self.mz_list)
         self.mz_sum = sum(self.mz_list)
         self.length_in_seconds = self.rt_list[-1] - self.rt_list[0]
[…]
