+++ vimms/BOMAS.py
@@ -8,13 +8,13 @@ from vimms.Box import *
 from vimms.Common import *
 from vimms.Controller import TopN_SmartRoiController, WeightedDEWController, TopN_RoiController, \
     NonOverlapController, IntensityNonOverlapController, TopNBoxRoiController, FlexibleNonOverlapController, \
-    FixedScansController, RoiBuilder, AgentBasedController, TopNController
+    FixedScansController, AgentBasedController, TopNController
 from vimms.DsDA import get_schedule, dsda_get_scan_params, create_dsda_schedule
 from vimms.Environment import *
 from vimms.Evaluation import evaluate_multi_peak_roi_aligner
 from vimms.Evaluation import evaluate_multiple_simulated_env
 from vimms.GridEstimator import *
-from vimms.Roi import FrequentistRoiAligner
+from vimms.Roi import FrequentistRoiAligner, RoiBuilder
 
 
 def run_coverage_evaluation(box_file, mzml_file, half_isolation_window):
