+++ vimms/Controller/roi.py
@@ -1,4 +1,3 @@
-import bisect
 from copy import deepcopy
 
 import numpy as np
@@ -9,156 +8,7 @@ from vimms.Common import ROI_EXCLUSION_DEW, ROI_EXCLUSION_WEIGHTED_DEW
 from vimms.Controller.topN import TopNController
 from vimms.Exclusion import MinIntensityFilter, LengthFilter, SmartROIFilter, WeightedDEWExclusion, DEWFilter, \
     WeightedDEWFilter
-from vimms.Roi import match, Roi, SmartRoi
-
-
-class RoiBuilder():
-    ROI_TYPE_NORMAL = 'roi'
-    ROI_TYPE_SMART = 'smart'
-
-    def __init__(self, mz_tol, rt_tol, min_roi_intensity, min_roi_length, min_roi_length_for_fragmentation=1,
-                 initial_length_seconds=5, reset_length_seconds=100,
-                 intensity_increase_factor=2, drop_perc=0.01, length_units="scans",
-                 roi_type=ROI_TYPE_NORMAL, grid=None, register_all_roi=False):
-
-        # ROI stuff
-        self.min_roi_intensity = min_roi_intensity
-        self.mz_tol = mz_tol
-        self.mz_units = 'ppm'
-        self.rt_tol = rt_tol
-        self.min_roi_length = min_roi_length
-        self.min_roi_length_for_fragmentation = min_roi_length_for_fragmentation
-        self.length_units = length_units
-        self.roi_type = roi_type
-        assert self.roi_type in [RoiBuilder.ROI_TYPE_NORMAL, RoiBuilder.ROI_TYPE_SMART]
-
-        # Create ROI
-        self.live_roi = []
-        self.dead_roi = []
-        self.junk_roi = []
-        self.live_roi_fragmented = []
-        self.live_roi_last_rt = []  # last fragmentation time of ROI
-
-        # fragmentation to Roi dictionaries
-        self.frag_roi_dicts = []  # scan_id, roi_id, precursor_intensity
-        self.roi_id_counter = 0
-
-        # Only used by SmartROI
-        self.initial_length_seconds = initial_length_seconds
-        self.reset_length_seconds = reset_length_seconds
-        self.intensity_increase_factor = intensity_increase_factor
-        self.drop_perc = drop_perc
-
-        # Grid stuff
-        self.grid = grid
-        self.register_all_roi = register_all_roi
-
-    def update_roi(self, new_scan):
-        if new_scan.ms_level == 1:
-            order = np.argsort(self.live_roi)
-            self.live_roi.sort()
-            self.live_roi_fragmented = np.array(self.live_roi_fragmented)[order].tolist()
-            self.live_roi_last_rt = np.array(self.live_roi_last_rt)[order].tolist()
-            current_ms1_scan_rt = new_scan.rt
-            not_grew = set(self.live_roi)
-            for idx in range(len(new_scan.intensities)):
-                intensity = new_scan.intensities[idx]
-                mz = new_scan.mzs[idx]
-                if intensity >= self.min_roi_intensity:
-                    roi = self._get_roi_obj(mz, 0, 0, None)
-                    match_roi = match(roi, self.live_roi, self.mz_tol, mz_units=self.mz_units)
-                    if match_roi:
-                        match_roi.add(mz, current_ms1_scan_rt, intensity)
-                        if match_roi in not_grew:
-                            not_grew.remove(match_roi)
-                    else:
-                        new_roi = self._get_roi_obj(mz, current_ms1_scan_rt, intensity, self.roi_id_counter)
-                        self.roi_id_counter += 1
-                        bisect.insort_right(self.live_roi, new_roi)
[…]
