### 2022-01-29 #3990 @MKoesters
Hi,

I'm very sorry, I started this in December and forgot about it due to an extremely busy January.
I'll do the rest of the review now, although all the formal requirements are met imo.

I did not check the box for the community guidelines, since they are not linked properly in the README or documentation.

Best,
Manuel
