+++ requirements.txt
@@ -1,166 +1,0 @@
-anyio==3.3.1
-appdirs==1.4.4
-argon2-cffi==21.1.0
-astunparse==1.6.3
-async-generator==1.10
-attrs==21.2.0
-Babel==2.9.1
-backcall==0.2.0
-black==19.10b0
-bleach==4.1.0
-build==0.7.0
-cached-property==1.5.2
-Cerberus==1.3.4
-certifi==2021.5.30
-cffi==1.14.6
-chardet==3.0.4
-charset-normalizer==2.0.6
-click==8.0.1
-colorama==0.4.4
-cryptography==3.4.8
-cycler==0.10.0
-Cython==0.29.24
-debugpy==1.4.3
-decorator==5.1.0
-defusedxml==0.7.1
-distlib==0.3.2
-docutils==0.17.1
-entrypoints==0.3
-Events==0.4
-flake8==3.9.2
-ghp-import==2.0.1
-GPy==1.10.0
-greenlet==1.1.1
-idna==3.2
-importlib-metadata==4.8.1
-iniconfig==1.1.1
-intervaltree==3.1.0
-ipykernel==6.4.1
-ipyparallel==7.0.1
-ipython==7.27.0
-ipython-genutils==0.2.0
-ipywidgets==7.5.1
-jedi==0.18.0
-jeepney==0.7.1
-Jinja2==3.0.1
-joblib==1.0.1
-json5==0.9.6
-jsonpickle==2.0.0
-jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client==7.0.3
-jupyter-console==6.1.0
-jupyter-core==4.8.1
-jupyter-server==1.11.0
-jupyterlab==3.1.12
-jupyterlab-pygments==0.1.2
-jupyterlab-server==2.8.1
-keyring==23.2.1
-kiwisolver==1.3.2
-loguru==0.5.3
-lxml==4.6.3
-Markdown==3.3.4
-MarkupSafe==2.0.1
-mass-spec-utils==0.0.12
-matplotlib==3.4.3
-matplotlib-inline==0.1.3
-mccabe==0.6.1
-mergedeep==1.3.4
-mistune==0.8.4
-mkdocs==1.2.2
-mkdocs-autorefs==0.2.1
-mkdocstrings==0.15.2
-molmass==2021.6.18
-more-itertools==8.4.0
-nbclassic==0.3.2
-nbclient==0.5.4
-nbconvert==6.1.0
-nbformat==5.1.3
-nest-asyncio==1.5.1
-networkx==2.6.3
-notebook==6.4.4
-numpy==1.21.2
-orderedmultidict==1.0.1
-outcome==1.0.1
-packaging==20.9
-pandas==1.3.3
-pandocfilters==1.5.0
-paramz==0.9.5
-parso==0.8.2
-pathspec==0.9.0
-patsy==0.5.1
-pbr==5.6.0
-pep517==0.11.0
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==8.3.2
-pip-shims==0.5.3
-pipenv-setup==3.1.1
-pipfile==0.0.2
-pkginfo==1.7.1
-plette==0.2.3
-pluggy==1.0.0
-prometheus-client==0.11.0
-prompt-toolkit==3.0.20
-psims==0.1.43
-psutil==5.8.0
-ptyprocess==0.7.0
-py==1.10.0
-pycodestyle==2.7.0
-pycparser==2.20
-pyDOE==0.3.8
-pyflakes==2.3.1
-Pygments==2.10.0
-pymdown-extensions==8.2
-pymzml==2.4.7
-pyparsing==2.4.7
[…]
