### 2022-02-01 #3990 @jspaezp
@csoneson sorry to bother you but ...
1. I still cannot edit the buttons
2. I have accepted the invitation in the past
3. When clicking the link to review the invitation I get the message ` Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account. `
