### 2022-02-01 #3990 @jspaezp
Just re-accepted and I do seem to have edit controls on the post.
This is my only non-private github account so I dont think that was the issue
