### 2022-02-02 6e4d6c9
WIP impl of box finder based on line-sweep.
