+++ vimms/Box.py
@@ -1,11 +1,13 @@
 import math
+import bisect
 import itertools
 import random
 from abc import abstractmethod
-from collections import defaultdict
+from collections import defaultdict, deque
 from decimal import Decimal
 from functools import reduce
 
+import intervaltree
 import GPy
 import numpy as np
 from mass_spec_utils.data_import.mzmine import PickedBox
@@ -22,7 +24,8 @@ class Point():
 
 
 class Box():
-    def __init__(self, x1, x2, y1, y2, parents=None, min_xwidth=0, min_ywidth=0, intensity=0):
+    def __init__(self, x1, x2, y1, y2, parents=None, min_xwidth=0, min_ywidth=0, intensity=0, id=None):
+        self.id = id
         self.pt1 = Point(min(x1, x2), min(y1, y2))
         self.pt2 = Point(max(x1, x2), max(y1, y2))
         self.parents = [self] if parents is None else parents
@@ -50,7 +53,7 @@ class Box():
 
     def copy(self, xshift=0, yshift=0):
         return type(self)(self.pt1.x + xshift, self.pt2.x + xshift, self.pt1.y + yshift, self.pt2.y + yshift,
-                          parents=self.parents, intensity=self.intensity)
+                          parents=self.parents, intensity=self.intensity, id=self.id)
 
     def shift(self, xshift=0, yshift=0):
         self.pt1.x += xshift
@@ -158,7 +161,15 @@ class Grid():
 
     @abstractmethod
     def register_box(self, box): pass
+    
+    def register_boxes(self, boxes):
+        for b in boxes: self.register_box(b)
+        
+    def set_active_boxes(self, *args):
+        pass
 
+    def clear(self):
+        self.__init__(self.min_rt, self.max_rt, self.rt_box_size, self.min_mz, self.max_mz, self.mz_box_size)
 
 class DictGrid(Grid):
     @staticmethod
@@ -212,9 +223,9 @@ class LocatorGrid(Grid):
     @staticmethod
     def dummy_non_overlap(box, other_boxes):
         return 1.0
-
+        
     @staticmethod
-    def splitting_non_overlap(box, other_boxes):
+    def non_overlap_boxes(box, other_boxes):
         new_boxes = [box]
         for b in other_boxes:  # filter boxes down via grid with large boxes for this loop + boxes could be potentially sorted by size (O(n) insert time in worst-case)?
             if (box.overlaps_with_box(b)):  # quickly exits any box not overlapping new box
@@ -227,9 +238,15 @@ class LocatorGrid(Grid):
                             updated_boxes.extend(split_boxes)
                         else:
                             updated_boxes.append(b2)
-                if (not updated_boxes): return 0.0
+                if (not updated_boxes): return []
                 new_boxes = updated_boxes
-        return sum(b.area() for b in new_boxes) / box.area()
+        return new_boxes
+
+    @staticmethod
+    def splitting_non_overlap(box, other_boxes):
+        boxes = LocatorGrid.non_overlap_boxes(box, other_boxes)
+        if(boxes == []): return 0.0
+        else: return sum(b.area() for b in boxes) / box.area()
 
     def non_overlap(self, box):
         return self.splitting_non_overlap(box, self.get_boxes(box))
@@ -342,7 +359,136 @@ class AllOverlapGrid(LocatorGrid):
                 binned.append([])
             binned[b.num_overlaps() - 1].append(b)
         return binned
+        
+class SweepGrid():
+    # aiming for [1011, 1874, 2297, 2514, 2635, 2721]
+    def __init__(self, track=True):
+        self.all_boxes, self.active, self.was_active, self.active_ends = [], [], [], []
+        self.active_intervals = intervaltree.IntervalTree()
[…]
