### 2022-02-02 043b5c5
Initial code tidy up of ROI Builder
