+++ vimms/Roi.py
@@ -16,6 +16,7 @@ from scipy.stats import pearsonr
 from vimms.Box import GenericBox
 from vimms.Chromatograms import EmpiricalChromatogram
 from vimms.Evaluation import *
+from vimms.MassSpec import Scan
 
 
 class Roi(object):
@@ -354,10 +355,10 @@ class RoiBuilder():
     ROI_TYPE_NORMAL = 'roi'
     ROI_TYPE_SMART = 'smart'
 
-    def __init__(self, mz_tol, rt_tol, min_roi_intensity, min_roi_length, min_roi_length_for_fragmentation=1,
+    def __init__(self, mz_tol, rt_tol, min_roi_intensity, min_roi_length,
                  initial_length_seconds=5, reset_length_seconds=100,
                  intensity_increase_factor=2, drop_perc=0.01, length_units="scans",
-                 roi_type=ROI_TYPE_NORMAL, grid=None, register_all_roi=False):
+                 roi_type=ROI_TYPE_NORMAL, grid=None):
         """
         Initialises an ROI Builder object.
         :param mz_tol: the m/z tolerance when matching a new point to existing ROIs
@@ -374,7 +375,6 @@ class RoiBuilder():
         :param length_units: the length unit (either in 'scans' or 'seconds')
         :param roi_type: the type of ROI object generated, either ROI_TYPE_NORMAL or ROI_TYPE_SMART
         :param grid: non-overlap grid, if any. Used to track overlapping fragmentation of ROI across samples.
-        :param register_all_roi: ???
         """
 
         # ROI stuff
@@ -383,7 +383,6 @@ class RoiBuilder():
         self.mz_units = 'ppm'
         self.rt_tol = rt_tol
         self.min_roi_length = min_roi_length
-        self.min_roi_length_for_fragmentation = min_roi_length_for_fragmentation  # FIXME: this parameter is unused in this class?!
         self.length_units = length_units
         self.roi_type = roi_type
         assert self.roi_type in [RoiBuilder.ROI_TYPE_NORMAL, RoiBuilder.ROI_TYPE_SMART]
@@ -409,7 +408,6 @@ class RoiBuilder():
 
         # Grid stuff
         self.grid = grid
-        self.register_all_roi = register_all_roi  # FIXME: this parameter is unused in this class?!
 
     def update_roi(self, new_scan):
         """
@@ -467,7 +465,7 @@ class RoiBuilder():
                         self.live_roi_last_rt.insert(self.live_roi.index(new_roi), None)
 
                         # If a grid object is provided, then update the grid too
-                        if self.register_all_roi and self.grid is not None:
+                        if self.grid is not None:
                             self.grid.register_roi(new_roi)
 
             # Separate the ROIs that have not been grown into dead or junk ROIs
@@ -552,7 +550,7 @@ class RoiBuilder():
                                     'precursor_intensity': intensity})
 
         # grid stuff
-        if self.register_all_roi is False and self.grid is not None:
+        if self.grid is not None:
             self.grid.register_roi(self.live_roi[i])
         self.live_roi[i].max_fragmentation_intensity = max(self.live_roi[i].max_fragmentation_intensity, intensity)
 
@@ -575,7 +573,7 @@ class RoiBuilder():
         """
         Returns all ROIs
         """
-        # TODO: check that the codes after the commented line is what we want
+        # TODO: check with Ross that the codes after the commented line is what we want
         # return self.live_roi + self.dead_roi
 
         # process all the remaining live ones - keeping only those that
@@ -602,7 +600,7 @@ class RoiAligner(object):
                  min_mz_width=0.01,
                  n_categories=1):
         """
-        TODO: complete this
+        TODO: add docstring comment
         Creates a new ROI aligner
[…]
