+++ vimms/ChemicalSamplers.py
@@ -254,7 +254,7 @@ class MZMLRTandIntensitySampler(RTAndIntensitySampler):
         self.rt_bins = [(k, k + 1) for k in rt_bins.keys()]
         self.rt_probs = [v / total_intensity for v in rt_bins.values()]
 
-        good, junk = make_roi(str(self.mzml_file_name), mz_tol=self.roi_params.mz_tol,
+        good = make_roi(str(self.mzml_file_name), mz_tol=self.roi_params.mz_tol,
                               mz_units=self.roi_params.mz_units, \
                               min_length=self.roi_params.min_length, min_intensity=self.roi_params.min_intensity, \
                               start_rt=self.roi_params.start_rt, stop_rt=self.roi_params.stop_rt,
@@ -333,7 +333,7 @@ class MZMLChromatogramSampler(ChromatogramSampler):
         self.good_rois = self._extract_rois()
 
     def _extract_rois(self):
-        good, junk = make_roi(str(self.mzml_file_name), mz_tol=self.roi_params.mz_tol,
+        good = make_roi(str(self.mzml_file_name), mz_tol=self.roi_params.mz_tol,
                               mz_units=self.roi_params.mz_units, \
                               min_length=self.roi_params.min_length, min_intensity=self.roi_params.min_intensity, \
                               start_rt=self.roi_params.start_rt, stop_rt=self.roi_params.stop_rt,
