+++ vimms/old_unused_experimental/PeakDetector.py
@@ -14,10 +14,10 @@ QCB_MZML2CHEMS_DICT = {'min_ms1_intensity': 1.75E5,
 
 
 def get_rois(mzml, min_roi_length, mzml2chems_dict=QCB_MZML2CHEMS_DICT):
-    good_roi, junk_roi = make_roi(mzml, mz_tol=mzml2chems_dict['mz_tol'], mz_units=mzml2chems_dict['mz_units'],
+    good_roi = make_roi(mzml, mz_tol=mzml2chems_dict['mz_tol'], mz_units=mzml2chems_dict['mz_units'],
                                   min_length=min_roi_length, min_intensity=mzml2chems_dict['min_intensity'],
                                   start_rt=mzml2chems_dict['start_rt'], stop_rt=mzml2chems_dict['stop_rt'])
-    return good_roi, junk_roi
+    return good_roi
 
 
 def mzml2classificationdata(mzmls, mzml_picked_peaks_files, min_roi_length=5, mzml2chems_dict=QCB_MZML2CHEMS_DICT,
