### 2022-02-04 389a1a5
TopN exclusion now done using interval trees, instead of looping.
