+++ vimms/Controller/topN.py
@@ -110,8 +110,7 @@ class TopNController(Controller):
         return new_tasks
 
     def update_state_after_scan(self, scan):
-        # update dynamic exclusion list after time has been increased
-        self.exclusion.cleanup(scan)
+        pass
 
 
 class ScanItem(object):
