+++ vimms/Exclusion.py
@@ -96,16 +96,17 @@ class BoxHolder(object):
                 hits.add(r.data)
         return hits
 
-    def check_point_2(self, mz, rt):
-        """
-        An alternative method that searches both trees
-        Might be faster if there are lots of rt ranges that 
-        can map to a particular mz value
-        """
-        mz_regions = self.boxes_mz.at(mz)
-        rt_regions = self.boxed_rt.at(rt)
-        inter = mz_regions.intersection(rt_regions)
-        return [r.data for r in inter]
+    # FIXME: this produces a different result from check_point, do not use
+    # def check_point_2(self, mz, rt):
+    #     """
+    #     An alternative method that searches both trees
+    #     Might be faster if there are lots of rt ranges that 
+    #     can map to a particular mz value
+    #     """
+    #     mz_regions = self.boxes_mz.at(mz)
+    #     rt_regions = self.boxes_rt.at(rt)
+    #     inter = mz_regions.intersection(rt_regions)
+    #     return [r.data for r in inter]
 
     def is_in_box(self, mz, rt):
         """
@@ -162,9 +163,10 @@ class BoxHolder(object):
 
 class TopNExclusion(object):
     def __init__(self, initial_exclusion_list=None):
-        self.exclusion_list = []
-        if initial_exclusion_list is not None:  # copy initial list, if provided
-            self.exclusion_list = list(initial_exclusion_list)
+        self.exclusion_list = BoxHolder()
+        if initial_exclusion_list is not None:  # add initial list, if provided
+            for initial in initial_exclusion_list:
+                self.exclusion_list.add_box(initial)
 
     def is_excluded(self, mz, rt):
         """
@@ -173,15 +175,13 @@ class TopNExclusion(object):
         :param rt: RT value
         :return: True if excluded (with weight 0.0), False otherwise (weight 1.0)
         """
-        # TODO: make this faster?
-        for x in self.exclusion_list:
-            exclude_mz = x.from_mz <= mz <= x.to_mz
-            exclude_rt = x.from_rt <= rt <= x.to_rt
-            if exclude_mz and exclude_rt:
-                logger.debug(
-                    'Excluded precursor ion mz {:.4f} rt {:.2f} because of {}'.format(mz, rt, x))
-                return True, 0.0
-        return False, 1.0
+        excluded = self.exclusion_list.is_in_box(mz, rt)
+        if excluded:
+            logger.debug(
+                'Excluded precursor ion mz {:.4f} rt {:.2f}'.format(mz, rt))
+            return True, 0.0
+        else:
+            return False, 1.0
 
     def update(self, current_scan, ms2_tasks):
         """
@@ -190,35 +190,17 @@ class TopNExclusion(object):
         :param ms2_tasks: scheduled ms2 tasks
         """
         rt = current_scan.rt
-        temp_exclusion_list = []
         for task in ms2_tasks:
             for precursor in task.get('precursor_mz'):
                 mz = precursor.precursor_mz
                 mz_tol = task.get(ScanParameters.DYNAMIC_EXCLUSION_MZ_TOL)
                 rt_tol = task.get(ScanParameters.DYNAMIC_EXCLUSION_RT_TOL)
                 x = self._get_exclusion_item(mz, rt, mz_tol, rt_tol)
+                self.exclusion_list.add_box(x)
                 logger.debug('Time {:.6f} Created dynamic temporary exclusion window mz ({}-{}) rt ({}-{})'.format(
                     rt,
                     x.from_mz, x.to_mz, x.from_rt, x.to_rt
                 ))
-                temp_exclusion_list.append(x)
-        self.exclusion_list.extend(temp_exclusion_list)
-
-    def cleanup(self, current_scan):
-        """
[…]
