### 2022-02-04 b517076
make_roi tidy up
