+++ vimms/Roi.py
@@ -15,6 +15,7 @@ from scipy.stats import pearsonr
 
 from vimms.Box import GenericBox
 from vimms.Chromatograms import EmpiricalChromatogram
+from vimms.Common import MZ_UNITS_PPM, MZ_UNITS_DA
 from vimms.Evaluation import *
 from vimms.MassSpec import Scan
 
@@ -327,8 +328,8 @@ class RoiParams(object):
     A parameter object that stores various settings required for ROIBuilder
     """
 
-    def __init__(self, mz_tol=10, min_length=10, min_intensity=50000, start_rt=0, stop_rt=10000000,
-                 length_units="scans"):
+    def __init__(self, mz_tol=10, min_length=1, min_intensity=0, start_rt=0, stop_rt=10000000,
+                 mz_units=MZ_UNITS_PPM, length_units="scans"):
         """
         Initialises an RoiParams object
         :param mz_tol: m/z  tolerance
@@ -342,6 +343,7 @@ class RoiParams(object):
         self.min_intensity = min_intensity
         self.start_rt = start_rt
         self.stop_rt = stop_rt
+        self.mz_units = mz_units
         self.length_units = length_units
 
 
@@ -353,7 +355,7 @@ class RoiBuilder():
     ROI_TYPE_NORMAL = 'roi'
     ROI_TYPE_SMART = 'smart'
 
-    def __init__(self, mz_tol, rt_tol, min_roi_intensity, min_roi_length,
+    def __init__(self, mz_tol, rt_tol, min_roi_intensity, min_roi_length, mz_units='ppm',
                  initial_length_seconds=5, reset_length_seconds=100,
                  intensity_increase_factor=2, drop_perc=0.01, length_units="scans",
                  roi_type=ROI_TYPE_NORMAL, grid=None):
@@ -363,9 +365,7 @@ class RoiBuilder():
         :param rt_tol: dynamic exclusion window (in seconds) before an ROI can be fragmented again
         :param min_roi_length: minimum length of ROI (in length_units) before an ROI is considered to be junk ROI
                                (and can be discarded)
-        :param min_roi_length_for_fragmentation: minimum length of ROI (in length_units) before this ROI
-                                                 is considered suitable for fragmentation.
-                                                 Note: the check isn't done in this class!
+        :param mz_units: m/z units, either in 'Da' or 'ppm'
         :param initial_length_seconds: SmartROI parameter
         :param reset_length_seconds: SmartROI parameter
         :param intensity_increase_factor: SmartROI parameter
@@ -378,11 +378,12 @@ class RoiBuilder():
         # ROI stuff
         self.min_roi_intensity = min_roi_intensity
         self.mz_tol = mz_tol
-        self.mz_units = 'ppm'
+        self.mz_units = mz_units
         self.rt_tol = rt_tol
         self.min_roi_length = min_roi_length
         self.length_units = length_units
         self.roi_type = roi_type
+        assert self.mz_units in [MZ_UNITS_PPM, MZ_UNITS_DA]
         assert self.roi_type in [RoiBuilder.ROI_TYPE_NORMAL, RoiBuilder.ROI_TYPE_SMART]
 
         # Create ROI
@@ -857,18 +858,11 @@ class FrequentistRoiAligner(RoiAligner):
 # Make the RoI from an input file
 # mz_units = Da for Daltons
 # mz_units = ppm for ppm
-def make_roi(input_file, mz_tol=0.001, min_length=10, min_intensity=50000, start_rt=0, stop_rt=10000000,
-             length_units="scans"):
+def make_roi(input_file, roi_params):
     """
     Make ROIs from an input file
[…]
