### 2022-02-07 #3990 @joewandy
Thank you both reviewers. It's certainly not an easy task going through our codebase, and on behalf of the whole ViMMS team, I'd like to say we really appreciate your effort.

@csoneson can I confirm that we could start working on addressing the issues raised by the reviewers? I'm not familiar with the procedure here, so thanks a lot!
