### 2022-02-07 5dab0b3
Added __repr__ to class RoiParams
