### 2022-02-08 2938496
Bugfix register_roi
