### 2022-02-09 6e6c04c
WeighedDEW exclusion fixes
