### 2022-02-09 99e96af
WeightedDEW exclusion fixes
