### 2022-02-09 47849c3
Moved constants to Common.py
