+++ vimms/Common.py
@@ -122,6 +122,7 @@ MZ_UNITS_DA = 'Da'
 ROI_TYPE_NORMAL = 'roi'
 ROI_TYPE_SMART = 'smart'
 
+
 ########################################################################################################################
 # Common classes
 ########################################################################################################################
