+++ vimms/Roi.py
@@ -5,19 +5,23 @@ Chemical objects for simulation input.
 """
 import bisect
 import sys
+from collections import Counter
 
+import numpy as np
 import pandas as pd
 import pylab as plt
 import pymzml
+import statsmodels.api as sm
 from loguru import logger
+from mass_spec_utils.data_import.mzmine import load_picked_boxes, map_boxes_to_scans
+from mass_spec_utils.data_import.mzml import MZMLFile
 from mass_spec_utils.data_processing.alignment import Peak, PeakSet
 from scipy.stats import pearsonr
-from pprint import pprint
 
 from vimms.Box import GenericBox
 from vimms.Chromatograms import EmpiricalChromatogram
 from vimms.Common import MZ_UNITS_PPM, MZ_UNITS_DA, ROI_TYPE_NORMAL, ROI_TYPE_SMART
-from vimms.Evaluation import *
+from vimms.Evaluation import load_peakonly_boxes, load_xcms_boxes, get_precursor_intensities
 from vimms.MassSpec import Scan
 
 
