### 2022-02-10 45b858c
Tidied up constants
