+++ environment.yml
@@ -1,6 +1,5 @@
 name: vimms
 channels:
-- defaults
 - conda-forge
 dependencies:
   - numpy
@@ -23,6 +22,7 @@ dependencies:
   - pytest
   - intervaltree
   - jupyterlab
+  - ipywidgets
   - gpy
   - mkdocs
   - mkdocstrings
