+++ vimms/Controller/roi.py
@@ -4,9 +4,11 @@ import numpy as np
 from loguru import logger
 from mass_spec_utils.data_import.mzml import MZMLFile
 
-from vimms.Common import ROI_EXCLUSION_DEW, ROI_EXCLUSION_WEIGHTED_DEW, MZ_UNITS_PPM, ROI_TYPE_NORMAL, ROI_TYPE_SMART
+from vimms.Common import ROI_EXCLUSION_DEW, ROI_EXCLUSION_WEIGHTED_DEW, \
+    MZ_UNITS_PPM, ROI_TYPE_NORMAL, ROI_TYPE_SMART
 from vimms.Controller.topN import TopNController
-from vimms.Exclusion import MinIntensityFilter, LengthFilter, SmartROIFilter, WeightedDEWExclusion, DEWFilter, \
+from vimms.Exclusion import MinIntensityFilter, LengthFilter, SmartROIFilter, \
+    WeightedDEWExclusion, DEWFilter, \
     WeightedDEWFilter
 from vimms.Roi import RoiBuilder
 
@@ -16,17 +18,26 @@ class RoiController(TopNController):
     An ROI based controller with multiple options
     """
 
-    def __init__(self, ionisation_mode, isolation_width, mz_tol, min_ms1_intensity, min_roi_intensity,
-                 min_roi_length, N, rt_tol=10, min_roi_length_for_fragmentation=1, length_units="scans", ms1_shift=0,
-                 params=None, exclusion_method=ROI_EXCLUSION_DEW, exclusion_t_0=None, mz_units=MZ_UNITS_PPM):
-        super().__init__(ionisation_mode, N, isolation_width, mz_tol, rt_tol, min_ms1_intensity, ms1_shift=ms1_shift,
+    def __init__(self, ionisation_mode, isolation_width, mz_tol,
+                 min_ms1_intensity, min_roi_intensity,
+                 min_roi_length, N, rt_tol=10,
+                 min_roi_length_for_fragmentation=1, length_units="scans",
+                 ms1_shift=0,
+                 params=None, exclusion_method=ROI_EXCLUSION_DEW,
+                 exclusion_t_0=None, mz_units=MZ_UNITS_PPM):
+        super().__init__(ionisation_mode, N, isolation_width, mz_tol, rt_tol,
+                         min_ms1_intensity, ms1_shift=ms1_shift,
                          params=params)
-        self.min_roi_length_for_fragmentation = min_roi_length_for_fragmentation
-        self.roi_builder = RoiBuilder(mz_tol, rt_tol, min_roi_intensity, min_roi_length,
-                                      length_units=length_units, roi_type=ROI_TYPE_NORMAL, mz_units=mz_units)
+        self.min_roi_length_for_fragmentation = min_roi_length_for_fragmentation # noqa
+        self.roi_builder = RoiBuilder(mz_tol, rt_tol, min_roi_intensity,
+                                      min_roi_length,
+                                      length_units=length_units,
+                                      roi_type=ROI_TYPE_NORMAL,
+                                      mz_units=mz_units)
 
         self.exclusion_method = exclusion_method
-        assert self.exclusion_method in [ROI_EXCLUSION_DEW, ROI_EXCLUSION_WEIGHTED_DEW]
+        assert self.exclusion_method in [ROI_EXCLUSION_DEW,
+                                         ROI_EXCLUSION_WEIGHTED_DEW]
         if self.exclusion_method == ROI_EXCLUSION_WEIGHTED_DEW:
             assert exclusion_t_0 is not None, 'Must be a number'
             assert exclusion_t_0 < rt_tol, 'Impossible combination'
@@ -38,7 +49,8 @@ class RoiController(TopNController):
         ms1_scan_params = self.get_ms1_scan_params()
[…]
