+++ vimms/old_unused_experimental/DataGenerator.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+
 import copy
 import glob
 import os
