+++ vimms/old_unused_experimental/PeakDetector.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+
 import numpy as np
 import pandas as pd
 

[…]
