+++ tests/test_controllers_TopN.py
@@ -3,9 +3,11 @@ import os
 import pymzml
 from loguru import logger
 
-from tests.conftest import N_CHEMS, MIN_MS1_INTENSITY, get_rt_bounds, CENTRE_RANGE, run_environment, \
+from tests.conftest import N_CHEMS, MIN_MS1_INTENSITY, get_rt_bounds, CENTRE_RANGE, \
+    run_environment, \
     check_non_empty_MS2, check_mzML, OUT_DIR, BEER_CHEMS, BEER_MIN_BOUND, BEER_MAX_BOUND, HMDB
-from vimms.ChemicalSamplers import EvenMZFormulaSampler, FixedMS2Sampler, UniformRTAndIntensitySampler, \
+from vimms.ChemicalSamplers import EvenMZFormulaSampler, FixedMS2Sampler, \
+    UniformRTAndIntensitySampler, \
     ConstantChromatogramSampler, DatabaseFormulaSampler
 from vimms.Chemicals import ChemicalMixtureCreator
 from vimms.Common import POSITIVE, set_log_level_warning, NEGATIVE, ScanParameters
@@ -70,7 +72,8 @@ class TestTopNForcedN:
 
 class TestTopNController:
     """
-    Tests the Top-N controller that does standard DDA Top-N fragmentation scans with the simulated mass spec class.
+    Tests the Top-N controller that does standard DDA Top-N fragmentation scans with the
+    simulated mass spec class.
     """
 
     def test_TopN_controller_with_simulated_chems(self, fragscan_dataset):
@@ -85,7 +88,8 @@ class TestTopNController:
 
         # create a simulated mass spec without noise and Top-N controller
         mass_spec = IndependentMassSpectrometer(ionisation_mode, fragscan_dataset)
-        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol, MIN_MS1_INTENSITY)
+        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol,
+                                    MIN_MS1_INTENSITY)
         min_bound, max_bound = get_rt_bounds(fragscan_dataset, CENTRE_RANGE)
 
         # create an environment to run both the mass spec and controller
@@ -111,9 +115,11 @@ class TestTopNController:
         # create a simulated mass spec with noise and Top-N controller
         mz_noise = GaussianPeakNoise(0.1)
         intensity_noise = GaussianPeakNoise(1000.)
-        mass_spec = IndependentMassSpectrometer(ionisation_mode, fragscan_dataset, mz_noise=mz_noise,
+        mass_spec = IndependentMassSpectrometer(ionisation_mode, fragscan_dataset,
+                                                mz_noise=mz_noise,
                                                 intensity_noise=intensity_noise)
-        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol, MIN_MS1_INTENSITY)
+        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol,
+                                    MIN_MS1_INTENSITY)
         min_bound, max_bound = get_rt_bounds(fragscan_dataset, CENTRE_RANGE)
 
         # create an environment to run both the mass spec and controller
@@ -138,7 +144,8 @@ class TestTopNController:
 
         # create a simulated mass spec without noise and Top-N controller
         mass_spec = IndependentMassSpectrometer(ionisation_mode, BEER_CHEMS)
-        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol, MIN_MS1_INTENSITY)
+        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol,
+                                    MIN_MS1_INTENSITY)
 
         # create an environment to run both the mass spec and controller
[…]
