+++ tests/test_controllers_agent.py
@@ -1,7 +1,7 @@
 from tests.conftest import OUT_DIR, check_mzML, check_non_empty_MS1, check_non_empty_MS2
 from vimms.Agent import TopNDEWAgent
-from vimms.ChemicalSamplers import UniformMZFormulaSampler, UniformRTAndIntensitySampler, GaussianChromatogramSampler, \
-    FixedMS2Sampler
+from vimms.ChemicalSamplers import UniformMZFormulaSampler, UniformRTAndIntensitySampler, \
+    GaussianChromatogramSampler, FixedMS2Sampler
 from vimms.Chemicals import ChemicalMixtureCreator
 from vimms.Common import POSITIVE, set_log_level_warning, set_log_level_debug
 from vimms.Controller import AgentBasedController
@@ -17,7 +17,8 @@ class TestAgentBasedController:
         ri = UniformRTAndIntensitySampler(min_rt=0, max_rt=80)
         cr = GaussianChromatogramSampler(sigma=1)
         ms = FixedMS2Sampler()
-        cs = ChemicalMixtureCreator(fs, rt_and_intensity_sampler=ri, chromatogram_sampler=cr, ms2_sampler=ms)
+        cs = ChemicalMixtureCreator(fs, rt_and_intensity_sampler=ri, chromatogram_sampler=cr,
+                                    ms2_sampler=ms)
         d = cs.sample(500, 2)
         ionisation_mode = POSITIVE
 
