+++ tests/test_mass_spec.py
@@ -1,7 +1,7 @@
 from loguru import logger
 
-from tests.conftest import MIN_MS1_INTENSITY, check_non_empty_MS2, check_mzML, OUT_DIR, BEER_CHEMS, BEER_MIN_BOUND, \
-    BEER_MAX_BOUND
+from tests.conftest import MIN_MS1_INTENSITY, check_non_empty_MS2, check_mzML, OUT_DIR, \
+    BEER_CHEMS, BEER_MIN_BOUND, BEER_MAX_BOUND
 from vimms.Common import POSITIVE
 from vimms.Controller import TopNController
 from vimms.Environment import Environment
@@ -10,7 +10,8 @@ from vimms.MassSpec import IndependentMassSpectrometer, TaskManager
 
 class TestSimulatedMassSpec:
     """
-    Tests the Top-N controller that does standard DDA Top-N fragmentation scans with the simulated mass spec class.
+    Tests the Top-N controller that does standard DDA Top-N fragmentation scans
+    with the simulated mass spec class.
     """
 
     def test_mass_spec(self):
@@ -23,8 +24,10 @@ class TestSimulatedMassSpec:
         ionisation_mode = POSITIVE
 
         task_manager = TaskManager(buffer_size=3)
-        mass_spec = IndependentMassSpectrometer(ionisation_mode, BEER_CHEMS, task_manager=task_manager)
-        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol, MIN_MS1_INTENSITY)
+        mass_spec = IndependentMassSpectrometer(ionisation_mode, BEER_CHEMS,
+                                                task_manager=task_manager)
+        controller = TopNController(ionisation_mode, N, isolation_width, mz_tol, rt_tol,
+                                    MIN_MS1_INTENSITY)
 
         # create an environment to run both the mass spec and controller
         env = Environment(mass_spec, controller, BEER_MIN_BOUND, BEER_MAX_BOUND, progress_bar=True)

[…]
