+++ .github/workflows/python-package.yml
@@ -30,8 +30,8 @@ jobs:
       run: |
         # stop the build if there are Python syntax errors or undefined names
         pipenv run flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --exclude=experimental/*
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        pipenv run flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        # exit-zero treats all errors as warnings.
+        pipenv run flake8 . --count --exit-zero --statistics
     - name: Test with pytest
       run: |
         pipenv run python -m pytest tests/
