### 2022-02-15 #3990 @jspaezp
@joewandy In my experience, we usually go over the comments and split them into what would be a "blocking issue" (things blocking the publication) and what would be a "review comment" (notes on the codebase that could be eventually addressed). This happens in a pretty organic way in my experience, just comment really liberally on the thread I started.

Then once the blocking issues have been addressed, we would give the thumbs up for the publication. In contrast with most other papers ... you can continuously discuss the comments raised with the reviewers (me ...) if you feel like some comment is unreasonable of if you want some clarification on any issue.
