### 2022-02-16 89f0cee
Box class reorganised internally to be more general geometry code. GridEstimator replaced with a BoxManager which takes responsibility for managing exclusion boxes, including splitting them for intensity non-overlap. Drift correction moved to its own file. BoxVisualise renamed to share the naming convention with the rest of the files.
