+++ vimms/Box.py
@@ -1,35 +1,79 @@
+import copy
 import math
-import bisect
 import itertools
-import random
-from abc import abstractmethod
-from collections import defaultdict, deque
+from abc import abstractmethod, ABCMeta
+from collections import defaultdict, namedtuple
 from decimal import Decimal
 from functools import reduce
+from operator import attrgetter
 
 import intervaltree
-import GPy
 import numpy as np
 from mass_spec_utils.data_import.mzmine import PickedBox
-from mass_spec_utils.library_matching.spectral_scoring_functions import cosine_similarity
-from mass_spec_utils.library_matching.spectrum import Spectrum
 
 
 class Point():
-    def __init__(self, x, y): self.x, self.y = float(x), float(y)
+    def __init__(self, x, y): 
+        self.x, self.y = float(x), float(y)
 
-    def __eq__(self, other_point): return math.isclose(self.x, other_point.x) and math.isclose(self.y, other_point.y)
+    def __eq__(self, other_point): 
+        return (
+            type(self) == type(other_point) 
+            and math.isclose(self.x, other_point.x) 
+            and math.isclose(self.y, other_point.y)
+        )
+        
+    def __hash__(self):
+        return (self.x, self.y).__hash__()
 
-    def __repr__(self): return "Point({}, {})".format(self.x, self.y)
+    def __repr__(self): 
+        return "Point({}, {})".format(self.x, self.y)
+    
+
+class Interval():
+    errmsg = "Interval has to be flat on one of the dimensions!"
+
+    def __init__(self, x1, x2, y1, y2):
+        assert math.isclose(x1, x2) or math.isclose(y1, y2), self.errmsg
+        self.pt1 = Point(min(x1, x2), min(y1, y2))
+        self.pt2 = Point(max(x1, x2), max(y1, y2))
+    
+    def __eq__(self, other_inv):
+        return (
+            type(self) == type(other_inv) 
+            and self.pt1 == other_inv.pt1 
+            and self.pt2 == other_inv.pt2
+        )
+        
+    def __hash__(self):
+        return (self.pt1, self.pt2).__hash__()
+
+    def __repr__(self):
+        return "Interval({}, {})".format(self.pt1, self.pt2)        
+    
+    def is_vertical(self):
+        return math.isclose(self.pt1.x, self.pt2.x)
+        
+    def is_horizontal(self):
+        return math.isclose(self.pt1.y, self.pt2.y)
 
 
 class Box():
-    def __init__(self, x1, x2, y1, y2, parents=None, min_xwidth=0, min_ywidth=0, intensity=0, id=None):
+    def __init__(self, x1, x2, y1, y2, 
+                    parents=None, 
+                    min_xwidth=0, 
+                    min_ywidth=0, 
+                    intensity=0, 
+                    id=None, 
+                    roi=None
+                ):
+                
         self.id = id
+        self.roi = roi
         self.pt1 = Point(min(x1, x2), min(y1, y2))
         self.pt2 = Point(max(x1, x2), max(y1, y2))
         self.parents = [self] if parents is None else parents
-        self.intensity = intensity
+        self.intensity = Decimal(intensity)
 
         if (self.pt2.x - self.pt1.x < min_xwidth):
             midpoint = self.pt1.x + ((self.pt2.x - self.pt1.x) / 2)
@@ -43,17 +87,29 @@ class Box():
         return "Box({}, {})".format(self.pt1, self.pt2)
 
     def __eq__(self, other_box):
-        return self.pt1 == other_box.pt1 and self.pt2 == other_box.pt2
+        return (
+            type(self) == type(other_box) 
+            and self.pt1 == other_box.pt1 
+            and self.pt2 == other_box.pt2
+        )
 
     def __hash__(self):
[…]
