+++ vimms/DriftCorrection.py
@@ -1,0 +1,169 @@
+from abc import abstractmethod
+
+import GPy
+
+from mass_spec_utils.library_matching.spectral_scoring_functions import cosine_similarity
+from mass_spec_utils.library_matching.spectrum import Spectrum
+
+class DriftModel():
+    @abstractmethod
+    def get_estimator(self, injection_number): pass
+
+    @abstractmethod
+    def _next_model(self): pass
+
+    def send_training_data(self, scan, roi, inj_num): pass
+
+    def send_training_pair(self, x, y): pass
+
+    def observed_points(self): return []
+
+    def update(self, **kwargs): pass
+
+
+class IdentityDrift(DriftModel):
+    '''Dummy drift model which does nothing, for testing purposes.'''
+
+    def get_estimator(self, injection_number): return lambda roi, inj_num: (0, {})
+
+    def _next_model(self, **kwargs): return self
+
+
+class OracleDrift(DriftModel):
+    '''Drift model that cheats by being given a 'true' rt drift fn. for every injection in simulation, for testing purposes.'''
+
+    def __init__(self, drift_fns):
+        self.drift_fns = drift_fns
+
+    def _next_model(self, **kwargs): return self
+
+    def get_estimator(self, injection_number):
+        if (type(self.drift_fns) == type([])): return self.drift_fns[injection_number]
+        return self.drift_fns
+
+
+class OraclePointMatcher():
+    MODE_ALLPOINTS = 0
+    MODE_RTENABLED = 1
+    MODE_FRAGPAIRS = 2
+
+    def __init__(self, chem_rts_by_injection, chemicals, max_points=None, mode=None):
+        if (not max_points is None and max_points < len(chem_rts_by_injection[0])):
+            idxes = random.sample([i for i, _ in enumerate(chem_rts_by_injection[0])], max_points)
+            self.chem_rts_by_injection = [[sample[i] for i in idxes] for sample in chem_rts_by_injection]
+        else:
+            self.chem_rts_by_injection = chem_rts_by_injection
+        self.chem_to_idx = {chem if chem.base_chemical is None else chem.base_chemical: idx for chem, idx in
+                            zip(chemicals, range(len(chem_rts_by_injection[0])))}
+        self.not_sent = [True] * len(self.chem_rts_by_injection[0])
+        self.available = [False] * len(self.chem_rts_by_injection[0])
+        self.mode = OraclePointMatcher.MODE_FRAGPAIRS if mode is None else mode
+
+    def _next_model(self):
+        self.not_sent = [True] * len(self.chem_rts_by_injection[0])
+
+    def send_training_data(self, model, scan, roi, inj_num):
+        if (self.mode == OraclePointMatcher.MODE_FRAGPAIRS):
+            if (not scan.fragevent is None):
+                parent_chem = scan.fragevent.chem if scan.fragevent.chem.base_chemical is None else scan.fragevent.chem.base_chemical
+                if (parent_chem in self.chem_to_idx):
+                    i = self.chem_to_idx[parent_chem]
+                    if (inj_num == 0):
+                        self.available[i] = True
+                    elif (self.available[i] and self.not_sent[i]):
+                        model.send_training_pair(self.chem_rts_by_injection[inj_num][i],
+                                                 self.chem_rts_by_injection[0][i])
+                        self.not_sent[i] = False
+        else:
+            if (self.mode == OraclePointMatcher.MODE_RTENABLED):
+                enable = lambda y: scan.rt > y
+            else:
+                enable = lambda y: True
+
+            for i, (y, x) in enumerate(zip(self.chem_rts_by_injection[inj_num], self.chem_rts_by_injection[0])):
+                if (self.not_sent[i] and enable(y)):
+                    model.send_training_pair(y, x)
[…]

[…]
