+++ vimms/Matching.py
@@ -31,9 +31,22 @@ class MatchingScan():
         self.mzs = mzs
         self.intensities = intensities
         
-    def __repr__(self): return f"MatchingScan(num_injection={self.num_injection}, ms_level={self.ms_level}, rt={self.rt})"
-    def __eq__(self, other): return isinstance(other, type(self)) and self.num_injection == other.num_injection and self.ms_level == other.ms_level and math.isclose(self.rt, other.rt)
-    def __hash__(self): return (self.num_injection, self.ms_level, self.rt).__hash__()
+    def __repr__(self): 
+        return (
+            f"MatchingScan(num_injection={self.num_injection}, "
+            f"ms_level={self.ms_level}, rt={self.rt})"
+        )
+        
+    def __eq__(self, other): 
+        return (
+            isinstance(other, type(self)) 
+            and self.num_injection == other.num_injection 
+            and self.ms_level == other.ms_level 
+            and math.isclose(self.rt, other.rt)
+        )
+        
+    def __hash__(self): 
+        return (self.num_injection, self.ms_level, self.rt).__hash__()
         
     def interpolate_scan(left, right, mz_window):
         left_mzs, left_intensities = zip(*left.peaks)
@@ -49,8 +62,10 @@ class MatchingScan():
         in_window = []
         left_bound, right_bound = 0, 1
         for i, intensity in enumerate(intensities):
-            while(mzs[i] - mzs[left_bound] > mz_window): left_bound += 1
-            while(right_bound + 1 < mzs.shape[0] and mzs[right_bound + 1] - mzs[i] < mz_window): right_bound += 1
+            while(mzs[i] - mzs[left_bound] > mz_window): 
+                left_bound += 1
+            while(right_bound + 1 < mzs.shape[0] and mzs[right_bound + 1] - mzs[i] < mz_window): 
+                right_bound += 1
             in_window.append((left_bound, right_bound+1))
         
         return left_rt, right_rt, mzs, intensities, owner, in_window
@@ -58,17 +73,28 @@ class MatchingScan():
     @staticmethod
     def create_scan_intensities(mzml_path, num_injection, schedule, mz_window):
         new_scans = []
-        original_scans = sorted((s for s in MZMLFile(mzml_path).scans if s.ms_level == 1), key=lambda s: s.rt_in_seconds, reverse=True)
+        
+        ms1s = (s for s in MZMLFile(mzml_path).scans if s.ms_level == 1)
+        original_scans = sorted(ms1s, key=lambda s: s.rt_in_seconds, reverse=True)
             
-        left_rt, right_rt, mzs, intensities, owner, in_window = MatchingScan.interpolate_scan(original_scans[-1], original_scans[-2], mz_window)
+        left_rt, right_rt, mzs, intensities, owner, in_window = (
+            MatchingScan.interpolate_scan(original_scans[-1], original_scans[-2], mz_window)
+        )
         
         for s in schedule:
             try: ms_level, rt = s.ms_level, s.rt
             except AttributeError: ms_level, rt = s
             
             if(len(original_scans) > 1 and original_scans[-2].rt_in_seconds < rt): 
-                while(len(original_scans) > 1 and original_scans[-2].rt_in_seconds < rt): original_scans.pop()
-                if(len(original_scans) > 1): left_rt, right_rt, mzs, intensities, owner, in_window = MatchingScan.interpolate_scan(original_scans[-1], original_scans[-2], mz_window)
+                while(len(original_scans) > 1 and original_scans[-2].rt_in_seconds < rt): 
+                    original_scans.pop()
+                if(len(original_scans) > 1): 
+                    left_rt, right_rt, mzs, intensities, owner, in_window = (
+                        MatchingScan.interpolate_scan(original_scans[-1], 
[…]
