+++ vimms/Box.py
@@ -97,7 +97,10 @@ class Box():
         return (self.pt1, self.pt2).__hash__()
 
     def area(self):
-        return (Decimal(self.pt2.x) - Decimal(self.pt1.x)) * (Decimal(self.pt2.y) - Decimal(self.pt1.y))
+        return (
+            (Decimal(self.pt2.x) - Decimal(self.pt1.x))
+            * (Decimal(self.pt2.y) - Decimal(self.pt1.y))
+        )
 
     def copy(self, xshift=0, yshift=0):
         return type(self)(
@@ -154,14 +157,22 @@ class GenericBox(Box):
 
     def overlap_2(self, other_box):
         if (not self.overlaps_with_box(other_box)): return 0.0
-        b = Box(max(self.pt1.x, other_box.pt1.x), min(self.pt2.x, other_box.pt2.x), max(self.pt1.y, other_box.pt1.y),
-                min(self.pt2.y, other_box.pt2.y))
+        b = type(self)(
+                max(self.pt1.x, other_box.pt1.x), 
+                min(self.pt2.x, other_box.pt2.x), 
+                max(self.pt1.y, other_box.pt1.y),
+                min(self.pt2.y, other_box.pt2.y)
+            )
         return b.area() / (self.area() + other_box.area() - b.area())
 
     def overlap_3(self, other_box):
         if (not self.overlaps_with_box(other_box)): return 0.0
-        b = Box(max(self.pt1.x, other_box.pt1.x), min(self.pt2.x, other_box.pt2.x), max(self.pt1.y, other_box.pt1.y),
-                min(self.pt2.y, other_box.pt2.y))
+        b = type(self)(
+                max(self.pt1.x, other_box.pt1.x), 
+                min(self.pt2.x, other_box.pt2.x), 
+                max(self.pt1.y, other_box.pt1.y),
+                min(self.pt2.y, other_box.pt2.y)
+            )
         return b.area() / self.area()
 
     def non_overlap_split(self, other_box):
@@ -175,24 +186,37 @@ class GenericBox(Box):
         split_boxes = []
         if (other_box.pt1.x > self.pt1.x):
             x1 = other_box.pt1.x
-            split_boxes.append(GenericBox(self.pt1.x, x1, y1, y2, parents=self.parents, intensity=self.intensity))
+            split_boxes.append(
+                GenericBox(self.pt1.x, x1, y1, y2, parents=self.parents, intensity=self.intensity)
+            )
         if (other_box.pt2.x < self.pt2.x):
             x2 = other_box.pt2.x
-            split_boxes.append(GenericBox(x2, self.pt2.x, y1, y2, parents=self.parents, intensity=self.intensity))
+            split_boxes.append(
+                GenericBox(x2, self.pt2.x, y1, y2, parents=self.parents, intensity=self.intensity)
+            )
         if (other_box.pt1.y > self.pt1.y):
             y1 = other_box.pt1.y
-            split_boxes.append(GenericBox(x1, x2, self.pt1.y, y1, parents=self.parents, intensity=self.intensity))
+            split_boxes.append(
+                GenericBox(x1, x2, self.pt1.y, y1, parents=self.parents, intensity=self.intensity)
+            )
         if (other_box.pt2.y < self.pt2.y):
             y2 = other_box.pt2.y
-            split_boxes.append(GenericBox(x1, x2, y2, self.pt2.y, parents=self.parents, intensity=self.intensity))
+            split_boxes.append(
+                GenericBox(x1, x2, y2, self.pt2.y, parents=self.parents, intensity=self.intensity)
+            )
         return split_boxes
 
     def split_all(self, other_box):
         if (not self.overlaps_with_box(other_box)): return None, None, None
         both_parents = self.parents + other_box.parents
[…]
