### 2022-02-16 3dc4b3c
Update README.md
added correct battleships talk link
