### 2022-02-21 4d28532
Fix for interval tree code accounting for the fact the package uses half-open intervals. Added a method to create a Scan object from an MZMLScan object. Some formatting changes.
