### 2022-02-21 df17e4c
Reimpl BoxIntervalTrees to exploit that overlaps are much more sparse on mz axis than rt
