### 2022-02-22 64fcfe7
Properly implemented abstract classes and methods using ABC. Also In Python 3, classes extend object implicitly.
