+++ vimms/MassSpec.py
@@ -10,7 +10,7 @@ from vimms.Common import adduct_transformation, DEFAULT_SCAN_TIME_DICT, \
 from vimms.Noise import NoPeakNoise
 
 
-class Peak(object):
+class Peak():
     """
     A class to represent an empirical or sampled scan-level peak object
     """
@@ -43,7 +43,7 @@ class Peak(object):
                self.ms_level == other.ms_level
 
 
-class Scan(object):
+class Scan():
     """
     A class to store scan information
     """
@@ -84,7 +84,7 @@ class Scan(object):
             self.scan_id, self.num_peaks, self.rt, self.ms_level)
 
 
-class ScanEvent(object):
+class ScanEvent():
     """
     A class to store fragmentation events. Mostly used for benchmarking purpose
     """
@@ -125,7 +125,7 @@ class ScanEvent(object):
             self.ms_level, self.chem, self.query_rt)
 
 
-class TaskManager(object):
+class TaskManager():
     """
     A class to track how many new tasks (scan commands) that we can send,
     given the buffer size of the mass spec.
@@ -223,7 +223,7 @@ class TaskManager(object):
         return len(self.pending_tasks)
 
 
-class IndependentMassSpectrometer(object):
+class IndependentMassSpectrometer():
     """
     A class that represents (synchronous) mass spectrometry process.
     Independent here refers to how the intensity of each peak in a scan is
