+++ vimms/ChemicalSamplers.py
@@ -2,6 +2,8 @@
 Sampling classes for ChemicalMixtureCreator
 """
 
+from abc import ABC, abstractmethod
+
 import numpy as np
 from loguru import logger
 from mass_spec_utils.data_import.mzml import MZMLFile
@@ -23,7 +25,7 @@ MIN_MZ_MS2 = DEFAULT_MSN_SCAN_WINDOW[0]
 # Formula samplers
 ###############################################################################
 
-class FormulaSampler(object):
+class FormulaSampler(ABC):
     """
     Base class for formula sampler
     """
@@ -32,8 +34,9 @@ class FormulaSampler(object):
         self.min_mz = min_mz
         self.max_mz = max_mz
 
+    @abstractmethod
     def sample(self, n_formulas):
-        raise NotImplementedError
+        pass
 
 
 class DatabaseFormulaSampler(FormulaSampler):
@@ -66,7 +69,7 @@ class DatabaseFormulaSampler(FormulaSampler):
         sub_formulas = list(
             filter(lambda x: Formula(x[0]).mass >= self.min_mz and Formula(
                 x[0]).mass <= self.max_mz - offset,
-                formulas))
+                   formulas))
         logger.debug(
             '{} unique formulas in filtered database'.format(
                 len(sub_formulas)))
@@ -94,7 +97,7 @@ class UniformMZFormulaSampler(FormulaSampler):
         :return: a list of Formula objects
         """
         mz_list = np.random.rand(n_formulas) * (
-            self.max_mz - self.min_mz) + self.min_mz
+                self.max_mz - self.min_mz) + self.min_mz
         return [(DummyFormula(m), None) for m in mz_list]
 
 
@@ -199,14 +202,15 @@ class MZMLFormulaSampler(FormulaSampler):
 ###############################################################################
 
 
-class RTAndIntensitySampler(object):
+class RTAndIntensitySampler(ABC):
     """
     Base class for RT and intensity sampler. Usually used when initialising
     a formula object.
     """
 
+    @abstractmethod
     def sample(self, formula):
-        raise NotImplementedError
+        pass
 
 
 class UniformRTAndIntensitySampler(RTAndIntensitySampler):
@@ -304,7 +308,7 @@ class MZMLRTandIntensitySampler(RTAndIntensitySampler):
                                              p=self.intensity_probs)
         intensity_bin = self.intensity_bins[intensity_bin_idx]
         log_intensity = np.random.rand() * (
-            intensity_bin[1] - intensity_bin[0]) + intensity_bin[0]
+                intensity_bin[1] - intensity_bin[0]) + intensity_bin[0]
         return rt, np.exp(log_intensity)
 
 
@@ -313,13 +317,14 @@ class MZMLRTandIntensitySampler(RTAndIntensitySampler):
 ###############################################################################
 
 
-class ChromatogramSampler(object):
+class ChromatogramSampler(ABC):
     """
     Base class for chromatogram sampler.
     """
 
+    @abstractmethod
     def sample(self, formula, rt, intensity):
-        raise NotImplementedError
+        pass
 
 
 class GaussianChromatogramSampler(ChromatogramSampler):
@@ -381,13 +386,14 @@ class MZMLChromatogramSampler(ChromatogramSampler):
 ###############################################################################
 
 
-class MS2Sampler(object):
+class MS2Sampler(ABC):
     """
     Base class for MS2 sampler
     """
 
+    @abstractmethod
     def sample(self, formula):
-        raise NotImplementedError
+        pass
 
 
 class UniformMS2Sampler(MS2Sampler):
@@ -431,8 +437,8 @@ class UniformMS2Sampler(MS2Sampler):
         s = sum(intensity_list)
[…]
