### 2022-02-22 d8d1093
Renamed sample_time() to sample() for consistency
