### 2022-02-22 22350f1
Fix for FixedScansController not implementing _process_scan()
