### 2022-02-22 4cbfe55
Tidied up chemical class hierarchy
