+++ vimms/Chemicals.py
@@ -3,6 +3,7 @@ Provides implementation of Chemicals objects that are used as input
 to the simulation.
 """
 import copy
+from abc import ABC, abstractmethod
 
 import numpy as np
 import scipy
@@ -144,54 +145,35 @@ class Adducts():
         return self.adduct_names
 
 
-class Chemical():
+class BaseChemical(ABC):
     """
-    The base class that represents a Chemical object.
+    The base class for Chemical objects across all MS levels.
+    Chemicals at MS level = 1 is special and should be instantiated as either Known
+    or Unknown chemicals.
+    For other MS levels, please use the MSN class.
+    """
+
+    def __init__(self, ms_level, children):
+        self.ms_level = ms_level
+        self.children = children
+
+
+class Chemical(BaseChemical):
+    """
+    The class that represents a Chemical object of MS-level 1.
     Should be realised as either Known or Unknown chemicals.
     """
 
-    def __repr__(self):
-        raise NotImplementedError()
+    def __init__(self, rt, max_intensity, chromatogram, children,
+                 base_chemical):
+        ms_level = 1
+        super().__init__(ms_level, children)
 
-
-class UnknownChemical(Chemical):
-    """
-    A Chemical representation from an unknown chemical formula
-    """
-
-    def __init__(self, mz, rt, max_intensity, chromatogram, children=None,
-                 base_chemical=None):
-        self.max_intensity = max_intensity
-        self.isotopes = [
-            (mz, 1, "Mono")]  # [(mz, intensity_proportion, isotope,name)]
-        self.adducts = {POSITIVE: [("M+H", 1)], NEGATIVE: [("M-H", 1)]}
         self.rt = rt
+        self.max_intensity = max_intensity
         self.chromatogram = chromatogram
-        self.children = children
-        self.ms_level = 1
         self.mz_diff = 0
-        self.mass = mz
         self.base_chemical = base_chemical
-
-    def __repr__(self):
-        return 'UnknownChemical mz=%.4f rt=%.2f max_intensity=%.2f' % (
-            self.isotopes[0][0], self.rt, self.max_intensity)
-
-    def get_key(self):
-        """
-        Turns a chemical object into (mz, rt, intensity) tuples for
-        equal comparison
-        :return: a tuple of the three values
-        """
-        return (tuple(self.isotopes), self.rt, self.max_intensity)
-
-    def __eq__(self, other):
-        if not isinstance(other, UnknownChemical):
-            return False
-        return self.get_key() == other.get_key()
-
-    def __hash__(self):
-        return hash(self.get_key())
 
     def get_apex_rt(self):
         return self.rt + self.chromatogram.get_apex_rt()
@@ -200,16 +182,63 @@ class UnknownChemical(Chemical):
         return self if self.base_chemical is None else \
             self.base_chemical.get_original_parent()
 
+    @abstractmethod
+    def get_key(self):
+        """
+        Turns a chemical object into some sensible keys for comparison
+        :return: keys for comparisons in __eq__ and __hash__
+        """
+        pass
+
+    def __eq__(self, other):
+        if type(other) is type(self):
+            return self.get_key() == other.get_key()
+        return False
+
+    def __hash__(self):
+        return hash(self.get_key())
+
+
+class UnknownChemical(Chemical):
+    """
+    A Chemical representation from an unknown chemical formula.
+    Unknown chemicals are typically created by extracting Regions-of-Interest
+    from an existing mzML file.
+    """
+
+    def __init__(self, mz, rt, max_intensity, chromatogram, children=None,
+                 base_chemical=None):
+        """
+        Initialises an UnknownChemical object.
+        """
+        super().__init__(rt, max_intensity, chromatogram, children, base_chemical)
[…]
