### 2022-02-23 f81d2db
Two new controllers added which do hard point exclusion on both saved DEW boxes and RoI boxes
