+++ vimms/Exclusion.py
@@ -75,6 +75,9 @@ class BoxHolder(object):
     def __init__(self):
         self.boxes_mz = IntervalTree()
         self.boxes_rt = IntervalTree()
+        
+    def __iter__(self):
+        return (inv.data for inv in self.boxes_rt.items())
 
     def add_box(self, box):
         """
