### 2022-02-23 f38ac1c
Implementing geometry operations on BoxGrid
