### 2022-02-23 b6da93c
Fix for box splitting and removed some test code.
