### 2022-02-24 f35e5dc
Initial notebook updates
