+++ examples/01. vimms (Wandy et al 2019)/DsDA/DsDA_Beer/DsDA_R_File.R
@@ -1,561 +1,0 @@
-#if (!requireNamespace("BiocManager", quietly = TRUE))
-# install.packages("BiocManager")
-#BiocManager::install("xcms", version = "3.8")
-
-
-library(xcms)
-library(mzR)
-
-DsDA<-function(
-  negmode=FALSE,
-  mzppm=20, 
-  pw=c(3,30), 
-  sn=20,
-  msscantime=0.18,
-  msmsscantime=0.035,
-  lockmassscantime=0.18,
-  isd=0.02, #not used
-  nmsms=4,
-  lockmassfrequency=20,
-  runtime=1020, # seem to need last time of schedule
-  lockmassrefine=FALSE,
-  fileIgnore='trash',
-  filterCond=c(2,20),
-  initLM=4.7,
-  msCE=6,
-  initCE=25,
-  altCE=7:60,
-  idleStop=10,
-  tofMult=16.625,
-  repeatMSMSfill=8,
-  schedule.file='DsDA_schedule.csv',
-  minPeaks=100,
-  min.rel.fragment.int=0.01,
-  min.abs.fragment.int=10,
-  init.CE.opt=10,
-  ionsLessThanPrecursorOnly=TRUE,
-  maximal.depth=NULL,
-  prefilter = c(20, 80),
-  mzdiff = 0.001,
-  dsda_dir=NA
-) {
-  
-  
-  
-  
-  #library(xcms)
-  #library(mzR)
-  
-  schedule<-read.csv(schedule.file, header=TRUE)
-  
-  ## having a run time as long as the method can cause WRENS problems, s
-  ## so we do not use the last 5 seconds of the run for WRENS  this is now handled within WRENS - more stable
-  # runtime<- runtime - 5
-  
-  schedule<-schedule[which(schedule$rt < (max(schedule$rt)-5)),]
-  headers<-c("MS1_MASS_DAC_SETTING",	"LM_RESOLUTION_SETTING",	"SOURCE_BIAS_SETTING",	
-             "SAMPLE_CONE_VOLTAGE_SETTING",	"targetMass",	"targetTime")
-  
-  if(any(msCE == altCE)) {
-    warning("CE ", msCE, " removed from altCE to ensure msCE is unique", '\n')
-    altCE <- altCE[!altCE %in% msCE]
-  }
-  
-  if(!is.null(maximal.depth)) {
-    depth.injections <- maximal.depth*(1:10000)
-  } else {
-    depth.injections <- NULL
-  }
-  inj.number <- 0
-  
-  setwd(dsda_dir)
-  lm<-which(schedule$type=="lm")
-  ms<-which(schedule$type=="ms")
-  msms<-which(schedule$type=="msms")
-  
-  #outtxt<-vector(length=nrow(schedule), mode="character")
-  # for(i in 1:length(schedule)) {
-  # outtxt[i]<-
-  
-  # }
-  
-  outcsv<-matrix(nrow=nrow(schedule), ncol=length(headers))
-  dimnames(outcsv)[[2]]<-headers
-  outcsv[,"targetTime"]<-schedule$rt
-  outcsv[,"SAMPLE_CONE_VOLTAGE_SETTING"]<- 30
-  outcsv[lm,"SOURCE_BIAS_SETTING"]<-NA
-  outcsv[ms,"SOURCE_BIAS_SETTING"]<-NA
-  outcsv[msms,"SOURCE_BIAS_SETTING"]<-NA
-  outcsv[msms,"LM_RESOLUTION_SETTING"]<-NA
-  cleanoutcsv<-outcsv
-  
-  ## assign random masses
[…]
