### 2022-02-25 296da95
Add more filtering based on intensity to RoiBuilder
