### 2022-02-27 389e744
Fixes to line sweep splitting boxes
