### 2022-02-28 e8725d7
Removed dependency on math.nextafter
