### 2022-02-28 d9a44c4
Allow certain scans to be skipped (up to max_skips_allowed) when building ROIs.
