+++ vimms/Roi.py
@@ -6,7 +6,7 @@ input.
 """
 import bisect
 import sys
-from collections import Counter
+from collections import Counter, defaultdict
 
 import numpy as np
 import pandas as pd
@@ -402,7 +402,8 @@ class RoiBuilder():
                  initial_length_seconds=5, reset_length_seconds=100,
                  intensity_increase_factor=2, drop_perc=0.01,
                  roi_type=ROI_TYPE_NORMAL, grid=None,
-                 min_roi_intensity_for_fragmentation=0):
+                 min_roi_intensity_for_fragmentation=0,
+                 max_skips_allowed=0):
         """
         Initialises an ROI Builder object.
         :param mz_tol: the m/z tolerance when matching a new point to
@@ -457,6 +458,11 @@ class RoiBuilder():
         # Grid stuff
         self.grid = grid
 
+        # Store the last few RT values of scans
+        self.max_skips_allowed = max_skips_allowed
+        self.rt_buffer = [0] * (self.max_skips_allowed+1) # 0 is current scan RT, others are past
+        self.skipped_roi_count = defaultdict(int)
+
     # flake8: noqa: C901
     def update_roi(self, new_scan):
         """
@@ -477,7 +483,25 @@ class RoiBuilder():
 
             # Current scan retention time of the MS1 scan is the RT of all
             # points in this scan
-            current_ms1_scan_rt = new_scan.rt
+            current_rt = new_scan.rt
+
+            # keep track of the last few RT values up to max_skips_allowed
+            # e.g. [110, 109, 107, 106, 103] for max_skips_allowed = 5
+            # the front of the list contains the current MS1 scan's RT value
+            self.rt_buffer.pop(-1)
+            self.rt_buffer.insert(0, current_rt)
+
+            # ensures data structure is in a consistent state
+            assert len(self.rt_buffer) == self.max_skips_allowed+1
+            assert self.rt_buffer[0] == current_rt
+
+            # check ordering, front element is larger than back
+            for i in range(len(self.rt_buffer) - 1):
+                assert self.rt_buffer[i] >= self.rt_buffer[i+1]
+
+            # check that there's no ROI with skip_count > self.max_skips_allowed
+            skip_counts = np.array(list(self.skipped_roi_count.values()))
+            assert np.sum(skip_counts > self.max_skips_allowed) == 0
 
             # The set of ROIs that are not grown yet.
             # Initially all currently live ROIs are included, and they're
@@ -486,32 +510,68 @@ class RoiBuilder():
 
             # For every (mz, intensity) in scan ..
             for idx in range(len(new_scan.intensities)):
-                intensity = new_scan.intensities[idx]
-                mz = new_scan.mzs[idx]
+                current_intensity = new_scan.intensities[idx]
+                current_mz = new_scan.mzs[idx]
 
-                if intensity >= self.min_roi_intensity:
+                if current_intensity >= self.min_roi_intensity:
 
                     # Create a dummy ROI object to represent the current m/z
                     # value. This produces either a normal ROI or smart ROI
                     # object, depending on self.roi_type
-                    roi = self._get_roi_obj(mz, 0, 0, None)
+                    roi = self._get_roi_obj(current_mz, 0, 0, None)
 
                     # Match dummy ROI to currently live ROIs based on mean
                     # m/z values. If no match, then return None
-                    match_roi = match(roi, self.live_roi, self.mz_tol)
-                    if match_roi:
+                    match_roi = self._match(roi, self.live_roi, self.mz_tol)
+                    if match_roi: # Got a match, so we grow this ROI
 
-                        # Got a match, so we grow this ROI
-                        match_roi.add(mz, current_ms1_scan_rt, intensity)
+                        # get how many scans have been skipped for this matched ROI
[…]
