### 2022-03-01 db7281c
Added tests for ROI Builder
