### 2022-03-02 a623a1a
- Renamed RoiParams to RoiBuilderParams - Introduced a new class SmartRoiParams - Simplify controller's constructors by passing these objects when possible
