+++ tests/test_controllers_roi.py
@@ -11,7 +11,7 @@ from vimms.Controller import TopN_RoiController, TopN_SmartRoiController, NonOve
 from vimms.Environment import Environment
 from vimms.GridEstimator import GridEstimator
 from vimms.MassSpec import IndependentMassSpectrometer, Scan
-from vimms.Roi import RoiBuilder
+from vimms.Roi import RoiBuilder, RoiBuilderParams, SmartRoiParams
 
 
 def delete_point(scans, mz, rt):
@@ -47,7 +47,8 @@ class TestRoiBuilder:
             scans.append(sc)
 
         # Build ROIs
-        roi_builder = RoiBuilder(10, 0, 0, 0)
+        roi_params = RoiBuilderParams()
+        roi_builder = RoiBuilder(roi_params)
         for sc in scans:
             roi_builder.update_roi(sc)
         rois = roi_builder.get_good_rois()
@@ -73,7 +74,8 @@ class TestRoiBuilder:
 
         # Build ROIs with no gap-filling.
         # We should get 4 ROIs as one ROI (at m/z 100) has been broken into two parts.
-        roi_builder = RoiBuilder(10, 0, 0, 0)
+        roi_params = RoiBuilderParams(max_gaps_allowed=0)
+        roi_builder = RoiBuilder(roi_params)
         for sc in scans:
             roi_builder.update_roi(sc)
         rois = roi_builder.get_good_rois()
@@ -81,7 +83,8 @@ class TestRoiBuilder:
 
         # Build ROIs with `max_skips_allowed` set to 1.
         # Here we still get 3 ROIs due to gap-filling.
-        roi_builder = RoiBuilder(10, 0, 0, 0, max_skips_allowed=1)
+        roi_params = RoiBuilderParams(max_gaps_allowed=1)
+        roi_builder = RoiBuilder(roi_params)
         for sc in scans:
             roi_builder.update_roi(sc)
         rois = roi_builder.get_good_rois()
@@ -97,7 +100,8 @@ class TestRoiBuilder:
         # Build ROIs with no gap-filling.
         # We should get 7 ROIs as the bottom ROI (at m/z 100) is now split into 4 parts, and
         # the top ROI (at m/z 102) is split into two with large gaps.
-        roi_builder = RoiBuilder(10, 0, 0, 0)
+        roi_params = RoiBuilderParams(max_gaps_allowed=0)
+        roi_builder = RoiBuilder(roi_params)
         for sc in scans:
             roi_builder.update_roi(sc)
         rois = roi_builder.get_good_rois()
@@ -105,7 +109,8 @@ class TestRoiBuilder:
 
         # Build ROIs with `max_skips_allowed` set to 3.
         # Here we still get 3 ROIs due to gap-filling.
-        roi_builder = RoiBuilder(10, 0, 0, 0, max_skips_allowed=3)
+        roi_params = RoiBuilderParams(max_gaps_allowed=3)
+        roi_builder = RoiBuilder(roi_params)
         for sc in scans:
             roi_builder.update_roi(sc)
         rois = roi_builder.get_good_rois()
@@ -135,9 +140,25 @@ class TestROIController:
 
         # create a simulated mass spec with noise and ROI controller
         mass_spec = IndependentMassSpectrometer(ionisation_mode, fragscan_dataset)
-        controller = TopN_RoiController(ionisation_mode, isolation_width, mz_tol,
-                                        MIN_MS1_INTENSITY, min_roi_intensity, min_roi_length,
-                                        N, rt_tol)
+
+        # def __init__(self,
+        #              ionisation_mode,
+        #              isolation_width,
+        #              N,
+        #              mz_tol,
+        #              rt_tol,
+        #              min_ms1_intensity,
[…]
