### 2022-03-02 ba1341a
Simplify controller parameters - Renamed RoiParams to RoiBuilderParams - Introduced a new class SmartRoiParams - Simplify controller's constructors by passing these objects when possible
