+++ vimms/Roi.py
@@ -234,9 +234,7 @@ class SmartRoi(Roi):
     AFTER_FRAGMENT = 2
     POST_PEAK = 3
 
-    def __init__(self, mz, rt, intensity, initial_length_seconds=5,
-                 reset_length_seconds=100,
-                 intensity_increase_factor=2, dew=15, drop_perc=0.01, id=None):
+    def __init__(self, mz, rt, intensity, smartroi_params, id=None):
         """
         Constructs a new Smart ROI
         :param mz: the initial m/z of this ROI.
@@ -245,20 +243,12 @@ class SmartRoi(Roi):
         Can be either a single value or a list of values.
         :param intensity: the initial intensity  of this ROI.
         Can be either a single value or a list of values.
-        :param initial_length_seconds: the initial length (in seconds) before
-        this ROI can be fragmented
-        :param reset_length_seconds: the length (in seconds) before this ROI
-        can be fragmented again (CAN_FRAGMENT)
-        :param intensity_increase_factor: a factor of which the intensity
-        should increase from the **minimum** since
-        last fragmentation before this ROI can be fragmented again
-        :param drop_perc: percentage drop in intensity since last fragmentation
-         before this ROI can be fragmented again
         :param id: the ID of this ROI
         """
         super().__init__(mz, rt, intensity, id=id)
+        self.params = smartroi_params
 
-        if initial_length_seconds > 0:
+        if self.params.initial_length_seconds > 0:
             self.status = SmartRoi.INITIAL_WAITING
             self.set_can_fragment(False)
         else:
@@ -267,12 +257,6 @@ class SmartRoi(Roi):
 
         self.min_frag_intensity = None
         self.intensity_diff = 0
-
-        self.initial_length_seconds = initial_length_seconds
-        self.reset_length_seconds = reset_length_seconds
-        self.intensity_increase_factor = intensity_increase_factor
-        self.drop_perc = drop_perc
-        self.dew = dew
 
     def fragmented(self):
         """
@@ -306,18 +290,18 @@ class SmartRoi(Roi):
         """
         super().add(mz, rt, intensity)
         if self.status == SmartRoi.INITIAL_WAITING:
-            if self.length_in_seconds >= self.initial_length_seconds:
+            if self.length_in_seconds >= self.params.initial_length_seconds:
                 self.status = SmartRoi.CAN_FRAGMENT
                 self.set_can_fragment(True)
         elif self.status == SmartRoi.AFTER_FRAGMENT:
             # in a period after a fragmentation has happened
             # if enough time has elapsed, reset everything
             if self.rt_list[-1] - self.rt_list[
-                self.fragmented_index] > self.reset_length_seconds:
+                self.fragmented_index] > self.params.reset_length_seconds:
                 self.status = SmartRoi.CAN_FRAGMENT
                 self.set_can_fragment(True)
             elif self.rt_list[-1] - self.rt_list[
-                self.fragmented_index] > self.dew:
+                self.fragmented_index] > self.params.dew:
                 # standard DEW has expired
                 # find the min intensity since the frag
                 # check current intensity -- if it is 5* when we fragmented,
@@ -325,10 +309,10 @@ class SmartRoi(Roi):
                 min_since_frag = min(
                     self.intensity_list[self.fragmented_index:])
                 if self.intensity_list[
-                    -1] > min_since_frag * self.intensity_increase_factor:
+                    -1] > min_since_frag * self.params.intensity_increase_factor:
                     self.status = SmartRoi.CAN_FRAGMENT
                     self.set_can_fragment(True)
-                elif self.intensity_list[-1] < self.drop_perc * \
+                elif self.intensity_list[-1] < self.params.drop_perc * \
[…]
