### 2022-03-03 740c2cb
JOSS tidy-up
Tidy up test fixtures
All experimental stuff are now moved to the `old_unused_experimental` branch
Remove unused test fixture
Got rid of a few flake8 problems
Format docstring for mkdoctrings correctly
Fixed typo in requirements
Improved API documentation
Notebook updates
