+++ vimms/old_unused_experimental/MultiSample.py
@@ -1,205 +1,0 @@
-# flake8: noqa
-
-from vimms.Box import IdentityDrift, LocatorGrid
-from vimms.Common import POSITIVE
-from vimms.Controller import TopNController, TopN_RoiController, TopNBoxRoiController, \
-    NonOverlapController
-from vimms.Environment import Environment
-from vimms.Evaluation import evaluate_multiple_simulated_env
-from vimms.GridEstimator import GridEstimator
-from vimms.MassSpec import IndependentMassSpectrometer
-from vimms.Roi import RoiAligner
-
-
-# import numpy as np
-#
-# from vimms.BOMAS import BaseOptimiser, QCB_MZML2CHEMS_DICT, XML_TEMPLATE_MS2, MZMINE_COMMAND
-# from vimms.Common import POSITIVE
-# from vimms.PythonMzmine import get_base_scoring_df, get_max_intensity
-#
-# QCB_ROI_CONTROLLER_PARAM_DICT = {"ionisation_mode": POSITIVE,
-#                                  "isolation_width": 1,
-#                                  "mz_tol": 10,
-#                                  "min_ms1_intensity": 1.75E5,
-#                                  "min_roi_intensity": 0,
-#                                  "min_roi_length": 0,
-#                                  "N": 10,
-#                                  "rt_tol": 10,
-#                                  "min_roi_length_for_fragmentation": 1,
-#                                  "rt_range": [(0, 1560)]}
-#
-#
-# QCB_XML_TEMPLATE_MS2 = 'C:\\Users\\Vinny\\work\\vimms\\batch_files\\QCB_mzmine_batch_ms2.xml'
-#
-# # TODO: set up a multisample environment?
-#
-#
-# class RepeatedExperiment(BaseOptimiser):
-#     def __init__(self, n_repeats, base_dir, ps,
-#                  # controller_method='Repeated_RoiController',
-#                  chem_param_dict=QCB_MZML2CHEMS_DICT,
-#                  controller_param_dict=QCB_ROI_CONTROLLER_PARAM_DICT,
-#                  # score_param_dict=QCB_SCORE_PARAM_DICT,
-#                  min_ms1_intensity=1.75E5,
-#                  # ms1_mzml=None,
-#                  # ms1_picked_peaks_file=None,
-#                  dataset_file=None,
-#                  add_noise=True,
-#                  # xml_template_ms1=XML_TEMPLATE_MS1,
-#                  xml_template_ms2=XML_TEMPLATE_MS2,
-#                  mzmine_command=MZMINE_COMMAND):
-#         super().__init__(base_dir, ps, 'Repeated_RoiController', chem_param_dict, controller_param_dict, None,
-#                          min_ms1_intensity,  None, None, dataset_file, add_noise, None,
-#                          xml_template_ms2, mzmine_command)
-#         self.n_repeats = n_repeats
-#         self.peaks_fragmented = [0]
-#         self.prop_peaks_fragmented = [0]
-#
-#     def run(self):
-#         next_flex_param_dict = {'peak_df': None}
-#         for i in range(self.n_repeats):
-#             env, controller_name = self._run_controller(self.mass_spec, self.controller_method,
-#                                                         self.controller_param_dict, next_flex_param_dict, i)
-#             if i == 0:
-#                 ms2_picked_peaks_file = self.frag_peak_picking(controller_name)
-#                 self.peak_df = get_base_scoring_df(ms2_picked_peaks_file)
-#                 self.peak_df['frag_status'] = [0 for j in range(len(self.peak_df.iloc[:, 0]))]
-#             # calculate peak frag status fo current iteration
-#             self.peak_df = self._update_peak_df(env.controller, self.peak_df)
[…]
