+++ vimms/old_unused_experimental/PlotsForPaper.py
@@ -1,728 +1,0 @@
-# flake8: noqa
-
-import os
-from collections import defaultdict
-
-import matplotlib.patches as mpatches
-import numpy as np
-import pandas as pd
-import pylab as plt
-import pymzml
-import seaborn as sns
-from loguru import logger
-
-from vimms.Chemicals import UnknownChemical
-from vimms.Common import load_obj, PROTON_MASS, find_nearest_index_in_array
-from vimms.MassSpec import ScanEvent
-from vimms.Roi import make_roi, RoiBuilderParams
-from vimms.old_unused_experimental.Chemicals import RoiToChemicalCreator
-from vimms.old_unused_experimental.SpectralUtils import get_precursor_info, get_chemicals
-
-
-def get_N(row):
-    if 'T10' in row['filename']:
-        return 10
-    else:
-        return row['filename'].split('_')[3]
-
-
-def get_dew(row):
-    if 'T10' in row['filename']:
-        return 15
-    else:
-        tok = row['filename'].split('_')[5]  # get the dew value in the filename
-        return tok.split('.')[0]  # get the part before '.mzML'
-
-
-def experiment_group(row):
-    if 'experiment' in row:
-        col_to_check = 'experiment'
-    else:
-        col_to_check = 'filename'
-
-    if 'beer' in row[col_to_check]:
-        return 'beer'
-    else:
-        return 'urine'
-
-
-def add_group_column(df):
-    df['group'] = df.apply(lambda row: experiment_group(row), axis=1)
-
-
-def get_df(csv_file, min_ms1_intensity, rt_range, mz_range):
-    df = pd.read_csv(csv_file)
-    return filter_df(df, min_ms1_intensity, rt_range, mz_range)
-
-
-def filter_df(df, min_ms1_intensity, rt_range, mz_range):
-    # filter by rt range
-    if rt_range is not None:
-        df = df[(df['rt'] > rt_range[0][0]) & (df['rt'] < rt_range[0][1])]
-
-    # filter by mz range
-    if mz_range is not None:
-        df = df[(df['rt'] > mz_range[0][0]) & (df['rt'] < mz_range[0][1])]
-
-    # filter by min intensity
-    intensity_col = 'maxo'
-    if min_ms1_intensity is not None:
-        df = df[(df[intensity_col] > min_ms1_intensity)]
-
-    # add log intensity column
-    df['log_intensity'] = df.apply(lambda row: np.log(row[intensity_col]), axis=1)
-
-    # add N column
-    try:
-        df['N'] = df.apply(lambda row: get_N(row), axis=1)
-        df[['N']] = df[['N']].astype('int')
-    except IndexError:
-        pass
-    except ValueError:
-        df['N'] = df.apply(lambda row: np.nan, axis=1)
-
-    # add group column
-    df['group'] = df.apply(lambda row: experiment_group(row), axis=1)
-    return df
-
-
-def make_boxplot(df, x, y, xticklabels, title, outfile=None):
-    g = sns.catplot(x=x, y=y, kind='box', data=df)
-    g.fig.set_size_inches(10, 3)
-    if xticklabels is not None:
-        g.set_xticklabels(xticklabels, rotation=90)
-    else:
-        g.set_xticklabels(rotation=90)
-    plt.title(title)
-    plt.tight_layout()
-    if outfile is not None:
-        plt.savefig(outfile, dpi=300)
-    plt.show()
-
-
-def make_hist(df, col_name, file_name, title):
-    gb = df.groupby('filename')
-    group_df = gb.get_group(file_name)
-    vals = group_df[col_name].values
-    logger.debug(vals, len(vals))
-    _ = plt.hist(vals, bins=100)
-    plt.title(title)
-    plt.tight_layout()
-    plt.show()
-
-
-def to_chemical(row):
[…]
