+++ vimms/Controller/abc.py
@@ -1,8 +1,24 @@
+"""
+This file contains the implementation of agent-based controllers.
+An agent is a useful abstraction that allows us to maintain
+global state outside a controller. This is useful when we need to track objects, e.g.
+precursors, fragmented counts, regions of interests etc, across injections.
+"""
+
 from vimms.Controller.base import Controller
 
 
 class AgentBasedController(Controller):
+    """
+    A class that implements an agent-based controller.
+    """
     def __init__(self, agent, advanced_params=None):
+        """Initialises an agent-based controller.
+
+        Arguments:
+            agent: an instance of the [vimms.Agent.AbstractAgent][] class.
+            advanced_params: optional advanced parameters for the mass spec.
+        """
         super().__init__(advanced_params=advanced_params)
         self.agent = agent
 
