### 2022-03-04 glasgowcompbio/vimms#230 @joewandy
Dear Juan (@jspaezp),

Thank you for your review. Your extensive comments have led to a large improvement in the software quality of the ViMMS project, and behalf of the entire team, I'd like to say that we really appreciate the effort.

The following is an itemised list of the issues found, and our comments to address them:

--

> 1. Contribution and authorship: Has the submitting author (@joewandy) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Our reply: I believe I've made a major contribution to the software 🙂 as you can see from the commit history. Other authors listed in the paper have also contributed appropriately as seen from the [Contributors page](https://github.com/glasgowcompbio/vimms/graphs/contributors) on Github.

--

> 2. Functionality: Have the functional claims of the software been confirmed?
> 3. Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Our reply: The functional claims of the software couldn't be verified because the example and demo notebooks were broken. We've since updated all notebooks to make sure they run.

--

> 4. Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The API documentation was indeed lacking. We've since improved the API documentation at https://vimms.readthedocs.io to be more comprehensive, including having better docstrings on the codes and parameters required. However since the project is still actively developed, some parts of the codes are not entirely documented yet, so this is an on-going work.

--

> 5. Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

A [README](https://github.com/glasgowcompbio/vimms/tree/master/tests/fixtures) has been added to explain the test fixtures and how they were generated. We'd look those files can be generated and cached during the testing process.

Coverage report has also been added to the CI, as seen [here](https://github.com/glasgowcompbio/vimms/runs/5423264281?check_suite_focus=true).

--

> 6. Community guidelines: Are there clear guidelines for third parties wishing to ...

We have now included a Contributing guideline [here](https://github.com/glasgowcompbio/vimms/blob/master/CONTRIBUTING.md), and made sure that it's now linked from the README. The Contributing guideline details the linter, testing and documentation framework used for anybody who wishes to contribute.

It took a while, but we've also fixed [all the flake8 violations](https://github.com/glasgowcompbio/vimms/runs/5423264281?check_suite_focus=true), as well as agreeing among ourselves not to introduce new ones in the future.

--

> 7. Code review:

- 1, 2, 3, 6, 9, 11 are fixed
- 4: thank you for the suggestion to import from `tqdm.auto`. The progress bar looks a lot nicer now.
- 7: `vimms/notebook` shouldn't be there and it has been moved out from the package itself
- 8, 9: we properly implement these classes as abstract classes now.
- 10: those experimental files have been moved to their own `old_unused_experimental ` branch.

--

I hope these changes are satisfactory to address the issues above. If there's any other problem, please let me know.

Thank you,

Joe Wandy (on behalf of the ViMMS team)
