### 2022-03-07 b6bc56d
Simplify score calculation (multiply only)
