### 2022-03-08 5c2de1b
Sweep-line fixed to use floats to 8 dec. pl. - since there is almost no arithmetic internally, we can just round off the imprecision that comes to us from outside in a 'closed system' - this is essentially like quantising the floats to larger bins. Edge-case with 'equal' boxes fixed, s.t. both boxes will correctly be used as parents of the combined box. Tidy-up of BoxManager components to be more modular. Small fix to InterpolationMap.
