+++ vimms/Box.py
@@ -13,8 +13,12 @@ import numpy as np
 from mass_spec_utils.data_import.mzmine import PickedBox
 
 class Point():
+    def round(self, ndigits=8):
+        self.x, self.y = round(self.x, ndigits), round(self.y, ndigits)
+
     def __init__(self, x, y): 
         self.x, self.y = float(x), float(y)
+        self.round()
 
     def __eq__(self, other_point): 
         return (
@@ -75,14 +79,17 @@ class Box():
         self.intensity = intensity
 
         if (self.pt2.x - self.pt1.x < min_xwidth):
-            midpoint = self.pt1.x + ((self.pt2.x - self.pt1.x) / 2)
-            self.pt1.x, self.pt2.x = midpoint - (min_xwidth / 2), midpoint + (
-                    min_xwidth / 2)
+            midpoint = self.pt1.x + (self.pt2.x - self.pt1.x) / 2
+            self.pt1.x = midpoint - min_xwidth / 2 
+            self.pt2.x = midpoint + min_xwidth / 2
 
         if (self.pt2.y - self.pt1.y < min_ywidth):
-            midpoint = self.pt1.y + ((self.pt2.y - self.pt1.y) / 2)
-            self.pt1.y, self.pt2.y = midpoint - (min_ywidth / 2), midpoint + (
-                    min_ywidth / 2)
+            midpoint = self.pt1.y + (self.pt2.y - self.pt1.y) / 2
+            self.pt1.y = midpoint - min_ywidth / 2 
+            self.pt2.y = midpoint + min_ywidth / 2
+            
+        self.pt1.round()
+        self.pt2.round()
 
     def __repr__(self):
         return "Box({}, {})".format(self.pt1, self.pt2)
@@ -120,6 +127,8 @@ class Box():
         self.pt2.x += xshift
         self.pt1.y += yshift
         self.pt2.y += yshift
+        self.pt1.round()
+        self.pt2.round()
 
     def num_overlaps(self):
         return len(self.parents)
@@ -212,22 +221,22 @@ class GenericBox(Box):
         if (other_box.pt1.x > self.pt1.x):
             x1 = other_box.pt1.x
             split_boxes.append(
-                GenericBox(self.pt1.x, x1, y1, y2, parents=self.parents, intensity=self.intensity)
+                type(self)(self.pt1.x, x1, y1, y2, parents=self.parents, intensity=self.intensity)
             )
         if (other_box.pt2.x < self.pt2.x):
             x2 = other_box.pt2.x
             split_boxes.append(
-                GenericBox(x2, self.pt2.x, y1, y2, parents=self.parents, intensity=self.intensity)
+                type(self)(x2, self.pt2.x, y1, y2, parents=self.parents, intensity=self.intensity)
             )
         if (other_box.pt1.y > self.pt1.y):
             y1 = other_box.pt1.y
             split_boxes.append(
-                GenericBox(x1, x2, self.pt1.y, y1, parents=self.parents, intensity=self.intensity)
+                type(self)(x1, x2, self.pt1.y, y1, parents=self.parents, intensity=self.intensity)
             )
         if (other_box.pt2.y < self.pt2.y):
             y2 = other_box.pt2.y
             split_boxes.append(
-                GenericBox(x1, x2, y2, self.pt2.y, parents=self.parents, intensity=self.intensity)
+                type(self)(x1, x2, y2, self.pt2.y, parents=self.parents, intensity=self.intensity)
             )
         return split_boxes
 
@@ -541,16 +550,7 @@ class BoxExact(BoxGeometry):
         box = box.copy()
         box.intensity = 0.0
         
-        #TODO: remove this
-        def box2pt(b): return (b.pt1.x, b.pt2.x, b.pt1.y, b.pt2.y)
[…]
