### 2022-03-09 1f663df
Write boxes to .csv (and back again!) for BatMass interop
