### 2022-03-16 5f2b530
Big rewrites to internals of intensity non-overlap, especially box splitting. All responsibility for turning RoIs to boxes moved to BoxManager, and min_mz_width and min_rt_width parameters taken with it. LineSweeper class split out from BoxExact subclass.
