+++ vimms/Box.py
@@ -112,10 +112,7 @@ class Box():
         return [self.min_rt, self.max_rt, self.min_mz, self.max_mz, self.intensity]
 
     def area(self):
-        return (
-            (self.pt2.x - self.pt1.x)
-            * (self.pt2.y - self.pt1.y)
-        )
+        return (self.pt2.x - self.pt1.x) * (self.pt2.y - self.pt1.y)
 
     def copy(self, xshift=0, yshift=0):
         return type(self)(
@@ -192,9 +189,17 @@ class GenericBox(Box):
                 and self.pt2.x >= other_box.pt2.x
                 and self.pt2.y >= other_box.pt2.y
         )
+        
+    def overlap_raw(self, other_box):
+        if(not self.overlaps_with_box(other_box)): 
+            return 0.0
+        return (
+            (min(self.pt2.x, other_box.pt2.x) - max(self.pt1.x, other_box.pt1.x))
+            * (min(self.pt2.y, other_box.pt2.y) - max(self.pt1.y, other_box.pt1.y))
+        )
 
     def overlap_2(self, other_box):
-        if (not self.overlaps_with_box(other_box)): 
+        if(not self.overlaps_with_box(other_box)): 
             return 0.0
         b = type(self)(
                 max(self.pt1.x, other_box.pt1.x), 
@@ -205,7 +210,7 @@ class GenericBox(Box):
         return b.area() / (self.area() + other_box.area() - b.area())
 
     def overlap_3(self, other_box):
-        if (not self.overlaps_with_box(other_box)): 
+        if(not self.overlaps_with_box(other_box)): 
             return 0.0
         b = type(self)(
                 max(self.pt1.x, other_box.pt1.x), 
@@ -388,8 +393,156 @@ class LocatorGrid(Grid):
             for s in row:
                 s.add(box)
                 
-    #TODO: probably can clear without reallocating array, because that's slow
+    def clear(self):
+        for i, row in enumerate(self.boxes):
+            for j, _ in enumerate(row):
+                self.boxes[i, j] = set()
 
+class LineSweeper():
+    EndPoint = namedtuple("EndPoint", ["pos", "is_end", "box"])
+    
+    def __init__(self):
+        self.all_boxes, self.active, self.was_active = [], [], []
+        self.active_intervals = intervaltree.IntervalTree()
+        self.previous_loc, self.current_loc = -1, 0
+        self.last_accessed = dict()
+        self.active_pointer = 0
+        self.removed = set()
+
+    def get_all_boxes(self):
+        return self.all_boxes
+        
+    def _add_active(self, new_ptr, current_loc):
+        for b in itertools.islice(self.all_boxes, self.active_pointer, new_ptr):
+            self.active.append(b)
+            self.active_intervals.addi(b.pt1.y, b.pt2.y + 1E-12, b)
+            self.last_accessed[b] = current_loc
+        self.active.sort(key=attrgetter("pt2.x"), reverse=True)
+        self.active_pointer = new_ptr
+    
+    def _remove_active(self):
+        b = self.active.pop()
+        self.was_active.append(b)
+        if(not b in self.removed):
+            self.active_intervals.removei(b.pt1.y, b.pt2.y + 1E-12, b)
+            self.removed.add(b)
+    
+    def set_active_boxes(self, current_loc):
+        if(current_loc < self.current_loc):
+            self.active_pointer = 0
+            self.active = []
+            self.active_intervals = intervaltree.IntervalTree()
+            self.previous_loc = -1
+            self.current_loc = 0
+            self.removed = set()
+            
+        new_ptr = self.active_pointer
+        while(new_ptr < len(self.all_boxes) and self.all_boxes[new_ptr].pt1.x <= current_loc):
[…]
