### 2022-03-18 9cf8b33
Grid controller topNEx fix
