### 2022-03-19 d84d498
Superfluous lines deleted and new intensity pointwise exclusion controller added
