### 2022-03-19 cf64484
Fix for smartroi intensity controllers using both smartroi and time filters
