### 2022-03-23 c7b911c
Code to attach info to peak boxes, and code to turn that info into an evaluation report
