+++ vimms/DDA_vs_DIA.py
@@ -1,0 +1,407 @@
+import glob
+import os
+
+import ipyparallel as ipp
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+import seaborn as sns
+from loguru import logger
+from mass_spec_utils.data_import.mzml import MZMLFile
+from mass_spec_utils.data_processing.mzmine import pick_peaks
+
+from vimms.Agent import TopNDEWAgent
+from vimms.Box import AllOverlapGrid, IdentityDrift, LocatorGrid
+from vimms.Box_Visualise import PlotPoints
+from vimms.Common import save_obj, load_obj
+from vimms.Controller import AgentBasedController, TopN_SmartRoiController, TopNController, \
+    WeightedDEWController, AIF, SWATH, AdvancedParams
+from vimms.Controller.box import IntensityNonOverlapController, NonOverlapController
+from vimms.Environment import Environment
+from vimms.Evaluation import evaluate_multi_peak_roi_aligner
+from vimms.GridEstimator import GridEstimator
+from vimms.MassSpec import IndependentMassSpectrometer
+from vimms.Roi import RoiAligner
+
+
+def run_experiment(result_folder, sample_list, controller_names, experiment_params):
+    """
+    Go through all the chemicals in result_folder, and run various controllers
+    on it. Will try to run it in parallel using ipcluster, if it fails then do it serially.
+
+    :param result_folder: the folder containing folders of chemicals in replicate
+    :param sample_list: the list of sample names to describe each replicate of chemical
+    :param controller_names: the list of controllers to run
+    :param experiment_params: experimental parameters
+    :return: None
+    """
+    pickle_files = glob.glob(os.path.join(result_folder, '*.p'))
+    params_list = []
+    for pf in pickle_files:
+        parameters = {
+            'pf': pf,
+            'sample_list': sample_list,
+            'result_folder': result_folder,
+            'controller_names': controller_names,
+            'experiment_params': experiment_params
+        }
+        params_list.append(parameters)
+
+    # Try to run the controllers in parallel. If fails, then run it serially
+    logger.warning('Running controllers in parallel, please wait ...')
+    run_serial = False
+    try:
+        rc = ipp.Client()
+        dview = rc[:]  # use all engines
+        with dview.sync_imports():
+            pass
+        dview.map_sync(run_once, params_list)
+    except OSError:  # cluster has not been started
+        run_serial = True
+    except ipp.error.TimeoutError:  # takes too long to run
+        run_serial = True
+
+    if run_serial:  # if any exception from above, try to run it serially
+        logger.warning(
+            'IPython cluster not found, running controllers in serial mode')
+        for parameters in params_list:
+            run_once(parameters)
+
+
+def run_once(parameters):
+    """
+    Run an experiment once
+    :param parameters: experimental parameters
+    :return: None
+    """
+    # get parameters
+    pf = parameters['pf']
+    sample_list = parameters['sample_list']
+    result_folder = parameters['result_folder']
+    controller_names = parameters['controller_names']
+    experiment_params = parameters['experiment_params']
+    pbar = False
+
+    # get experiment name
+    root, ext = os.path.splitext(pf)
+    experiment_name = root.split(os.sep)[-1]
+    logger.warning(pf)
+
+    # get chemical list for the experiment
+    chem_list = load_obj(pf)
+    assert len(chem_list) == len(sample_list)
+
+    # run a lot of controllers on the chemical list
[…]
