### 2022-03-28 51e86b5
Fixes for unused names in functions that weren't run
