### 2022-03-29 #3990 @csoneson
@joewandy I went through the submission and overall this looks great. Just a few minor details in the paper:

* Could you spell out m/z before the first time it is used?
* On line 88 it looks like there may be an "a" too much.
* Line 139: Github -> GitHub
* Line 172: java -> Java
* Line 172: fisher -> Fisher. Also, are the quotes needed here?
* Lines 154/176: J.J.J. van der Hooft's name is rendered differently - could you make these consistent?

You can generate a new proof with `@editorialbot generate pdf`, as above.

The next steps then are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Archive the release in Zenodo or a similar service
- [ ] Check that the archive has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
