### 2022-03-29 89a7402
Change Flexible Non-Overlap to be runnable (and hence testable)
