### 2022-03-29 9d6c0f4
Minor fixes to JOSS paper. Bumped version for new release.
