+++ paper/paper.bib
@@ -43,7 +43,7 @@
 
 @article{davies21_smartroi,
   title={Rapid Development of Improved Data-Dependent Acquisition Strategies},
-  author={Davies, Vinny and Wandy, Joe and Weidt, Stefan and Van Der Hooft, Justin JJ and Miller, Alice and Daly, R{\'o}n{\'a}n and Rogers, Simon},
+  author={Davies, Vinny and Wandy, Joe and Weidt, Stefan and J J van der Hooft, Justin and Miller, Alice and Daly, R{\'o}n{\'a}n and Rogers, Simon},
   journal={Analytical chemistry},
   volume={93},
   number={14},
@@ -54,7 +54,7 @@
 }
 
 @article{smith2015jamss,
-  title={JAMSS: proteomics mass spectrometry simulation in Java},
+  title={{JAMSS}: proteomics mass spectrometry simulation in {Java}},
   author={Smith, Rob and Prince, John T},
   journal={Bioinformatics},
   volume={31},
@@ -66,7 +66,7 @@
 }
 
 @article{noyce2013mspire,
-  title={Mspire-Simulator: LC-MS shotgun proteomic simulator for creating realistic gold standard data},
+  title={{Mspire-Simulator}: LC-MS shotgun proteomic simulator for creating realistic gold standard data},
   author={Noyce, Andrew B and Smith, Rob and Dalgleish, James and Taylor, Ryan M and Erb, KC and Okuda, Nozomu and Prince, John T},
   journal={Journal of proteome research},
   volume={12},
@@ -78,7 +78,7 @@
 }
 
 @article{goldfarb2016msacquisitionsimulator,
-  title={MSAcquisitionSimulator: data-dependent acquisition simulator for LC-MS shotgun proteomics},
+  title={{MSAcquisitionSimulator}: data-dependent acquisition simulator for LC-MS shotgun proteomics},
   author={Goldfarb, Dennis and Wang, Wei and Major, Michael B},
   journal={Bioinformatics},
   volume={32},
@@ -90,7 +90,7 @@
 }
 
 @article{kosters2021smiter,
-  title={SMITER—A Python Library for the Simulation of LC-MS/MS Experiments},
+  title={{SMITER}—A {Python} Library for the Simulation of LC-MS/MS Experiments},
   author={K{\"o}sters, Manuel and Leufken, Johannes and Leidel, Sebastian A},
   journal={Genes},
   volume={12},
@@ -102,7 +102,7 @@
 }
 
 @article{wang2015openms,
-  title={OpenMS-Simulator: an open-source software for theoretical tandem mass spectrum prediction},
+  title={{OpenMS-Simulator}: an open-source software for theoretical tandem mass spectrum prediction},
   author={Wang, Yaojun and Yang, Fei and Wu, Peng and Bu, Dongbo and Sun, Shiwei},
   journal={BMC bioinformatics},
   volume={16},
@@ -114,9 +114,9 @@
 }
 
 @MISC{Thermo_Fisher_Scientific_undated-ny,
-  title        = "Thermo Fisher Application Programming Interface",
+  title        = "{Thermo} {Fisher} {Application} {Programming} {Interface}",
   booktitle    = "Thermo Fisher Application Programming Interface",
-  author       = "{``Thermo Fisher Scientific''}",
+  author       = "{Thermo Fisher Scientific}",
   howpublished = "\url{https://github.com/thermofisherlsms/iapi}",
   note         = "Accessed: 2020-8-18"
 }
