### 2022-03-29 #3990 @joewandy
Thank you @csoneson. I've done all the following steps:

> * [X]  Make a tagged release of the software, and list the version tag of the archived version here.

The tagged release is here: https://github.com/glasgowcompbio/vimms/releases/tag/V2.0.5.

> * [X]  Archive the release in Zenodo or a similar service
> * [X]  Check that the archive has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
> * [X]  Please list the DOI of the archived version here.

Zenodo archive can be found here: https://zenodo.org/record/6395052. I've checked that the metadata is correct. The DOI of the archived version is 10.5281/zenodo.6395052.
