### 2022-03-29 f062122
Bump jupyter-server from 1.11.0 to 1.15.4
Bumps [jupyter-server](https://github.com/jupyter/jupyter_server) from 1.11.0 to 1.15.4.
- [Release notes](https://github.com/jupyter/jupyter_server/releases)
- [Changelog](https://github.com/jupyter-server/jupyter_server/blob/main/CHANGELOG.md)
- [Commits](https://github.com/jupyter/jupyter_server/compare/v1.11.0...v1.15.4)

---
updated-dependencies:
- dependency-name: jupyter-server
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
