+++ requirements.txt
@@ -51,7 +51,7 @@ jupyter==1.0.0
 jupyter-client==7.0.3
 jupyter-console==6.1.0
 jupyter-core==4.8.1
-jupyter-server==1.11.0
+jupyter-server==1.15.4
 jupyterlab==3.1.12
 jupyterlab-pygments==0.1.2
 jupyterlab-server==2.8.1
